### 2021-06-01 5944dc2
Hotfix affiliations in paper.md
