### 2021-06-03 9751917
Hotfix behavior of get_bmu
