### 2021-06-24 bc413ba
Implement DVFA, add tests, bump version
