+++ examples/art/dvfa_supervised.jl
@@ -8,6 +8,9 @@ Description:
 using Logging
 using DelimitedFiles
 using AdaptiveResonance
+using Random
+
+Random.seed!(0)
 
 # Set the log level
 LogLevel(Logging.Info)
@@ -19,7 +22,7 @@ include("../../test/test_utils.jl")
 data = load_iris("data/Iris.csv")
 
 # Train and classify
-art = DDVFA()
+art = DVFA()
 y_hat_train = train!(art, data.train_x, y=data.train_y)
 y_hat = classify(art, data.test_x)
 y_hat_bmu = classify(art, data.test_x, get_bmu=true)
@@ -29,6 +32,6 @@ perf_train = performance(y_hat_train, data.train_y)
 perf_test = performance(y_hat, data.test_y)
 perf_test_bmu = performance(y_hat_bmu, data.test_y)
 
-@info "DDVFA Training Perf: $perf_train"
-@info "DDVFA Testing Perf: $perf_test"
-@info "DDVFA Testing BMU Perf: $perf_test_bmu"
+@info "DVFA Training Perf: $perf_train"
+@info "DVFA Testing Perf: $perf_test"
+@info "DVFA Testing BMU Perf: $perf_test_bmu"
