+++ src/ART/DVFA.jl
@@ -1,0 +1,399 @@
+using Base: stop_reading
+"""
+    DVFA.jl
+
+Description:
+    Includes all of the structures and logic for running a Dual-Vigilance Fuzzy ART (DVFA) module.
+
+Authors:
+    MATLAB implementation: Leonardo Enzo Brito da Silva
+    Julia port: Sasha Petrenko <sap625@mst.edu>
+
+References:
+[1] L. E. Brito da Silva, I. Elnabarawy and D. C. Wunsch II, "Dual
+Vigilance Fuzzy ART," Neural Networks Letters. To appear.
+[2] G. Carpenter, S. Grossberg, and D. Rosen, "Fuzzy ART: Fast
+stable learning and categorization of analog patterns by an adaptive
+resonance system," Neural Networks, vol. 4, no. 6, pp. 759–771, 1991.
+Code written by Leonardo Enzo Brito da Silva
+Under the supervision of Dr. Donald C. Wunsch II
+"""
+
+"""
+    opts_DVFA()
+
+Dual Vigilance Fuzzy ART options struct.
+
+# Examples
+```julia-repl
+julia> my_opts = opts_DVFA()
+```
+"""
+@with_kw mutable struct opts_DVFA <: AbstractARTOpts @deftype Float64
+    # Lower-bound vigilance parameter: [0, 1]
+    rho_lb = 0.55; @assert rho_lb >= 0 && rho_lb <= 1
+    # Upper bound vigilance parameter: [0, 1]
+    rho_ub = 0.6; @assert rho_ub >= 0 && rho_ub <= 1
+    # Choice parameter: alpha > 0
+    alpha = 1e-3; @assert alpha > 0
+    # Learning parameter: (0, 1]
+    beta = 1; @assert beta > 0 && beta <= 1
+    # Display flag
+    display::Bool = true
+
+    max_epochs = 1
+end # opts_DVFA
+
+"""
+    DVFA <: AbstractART
+
+Dual Vigilance Fuzzy ARTMAP module struct.
+
+# Examples
+```julia-repl
+julia> DVFA()
+DVFA
+    opts: opts_DVFA
+    ...
+```
+"""
+mutable struct DVFA <: AbstractART
+    # Get parameters
+    opts::opts_DVFA
+    config::DataConfig
+
+    # Working variables
+    labels::Array{Int, 1}
+    W::Array{Float64, 2}
+    T::Array{Float64, 1}
+    M::Array{Float64, 1}
+    W_old::Array{Float64, 2}
+    map::Array{Int, 1}
+    n_categories::Int
+    n_clusters::Int
+    epoch::Int
+end # DVFA
+
+"""
+    DVFA()
+
+Implements a DVFA learner with default options.
+
+# Examples
+```julia-repl
+julia> DVFA()
+DVFA
+    opts: opts_DDVFA
+    ...
+```
+"""
+function DVFA()
+    opts = opts_DVFA()
+    DVFA(opts)
+end # DVFA()
+
+"""
+    DVFA(opts::opts_DVFA)
+
+Implements a DVFA learner with specified options.
+
+# Examples
+```julia-repl
+julia> my_opts = opts_DVFA()
+julia> DVFA(my_opts)
+DVFA
+    opts: opts_DVFA
+    ...
+```
+"""
+function DVFA(opts::opts_DVFA)
+    DVFA(
+        opts,                           # opts
+        DataConfig(),                   # config
+        Array{Int}(undef, 0),           # labels
+        Array{Float64}(undef, 0, 0),    # W
+        Array{Float64}(undef, 0),       # M
+        Array{Float64}(undef, 0),       # T
+        Array{Float64}(undef, 0, 0),    # W_old
+        Array{Float64}(undef, 0),       # map
+        0,                              # n_categories
+        0,                              # n_clusters
[…]

[…]
