### 2021-07-01 444b5bb
Define explicit abstract numerical types, bump
