+++ src/common.jl
@@ -1,3 +1,36 @@
+
+# -------------------------------------------
+# Document: common.jl
+# Author: Sasha Petrenko <sap625@mst.edu>
+# Description:
+#   Types and functions that are used throughout AdaptiveResonance.jl.
+# -------------------------------------------
+
+# -------------------------------------------
+# Aliases
+# -------------------------------------------
+#   **Taken from StatsBase.jl**
+#
+#  These types signficantly reduces the need of using
+#  type parameters in functions (which are often just
+#  for the purpose of restricting the arrays to real)
+#
+# These could be removed when the Base supports
+# covariant type notation, i.e. AbstractVector{<:Real}
+
+# Real-numbered aliases
+const RealArray{T<:Real, N} = AbstractArray{T, N}
+const RealVector{T<:Real} = AbstractArray{T, 1}
+const RealMatrix{T<:Real} = AbstractArray{T, 2}
+
+# Integered aliases
+const IntegerArray{T<:Integer, N} = AbstractArray{T, N}
+const IntegerVector{T<:Integer} = AbstractArray{T, 1}
+const IntegerMatrix{T<:Integer} = AbstractArray{T, 2}
+
+# Specifically floating-point aliases
+const RealFP = Union{Float32, Float64}
+
 """
     DataConfig
 
@@ -5,10 +38,10 @@ Conatiner to standardize training/testing data configuration.
 """
 mutable struct DataConfig
     setup::Bool
-    mins::Array{Float64, 1}
-    maxs::Array{Float64, 1}
-    dim::Int64
-    dim_comp::Int64
+    mins::RealVector
+    maxs::RealVector
+    dim::Integer
+    dim_comp::Integer
 end # DataConfig
 
 """
@@ -27,13 +60,13 @@ function DataConfig()
 end # DataConfig()
 
 """
-    DataConfig(mins::Array, maxs::Array)
+    DataConfig(mins::RealVector, maxs::RealVector)
 
 Convenience constructor for DataConfig, requiring only mins and maxs of the features.
 
 This constructor is used when the mins and maxs differ across features. The dimension is inferred by the length of the mins and maxs.
 """
-function DataConfig(mins::Array, maxs::Array)
+function DataConfig(mins::RealVector, maxs::RealVector)
     # Verify that the mins and maxs are the same length
     length(mins) != length(maxs) && error("Mins and maxs must be the same length.")
     # Get the dimension from one of the arrays
@@ -46,16 +79,16 @@ function DataConfig(mins::Array, maxs::Array)
         dim,    # dim
         dim*2   # dim_comp
     )
-end # DataConfig(mins::Array, maxs::Array)
+end # DataConfig(mins::RealVector, maxs::RealVector)
 
 """
-    DataConfig(min::Real, max::Real, dim::Int64)
+    DataConfig(min::Real, max::Real, dim::Integer)
 
 Convenience constructor for DataConfig, requiring only a global min, max, and dim.
 
 This constructor is used in the case that the feature mins and maxs are all the same respectively.
 """
-function DataConfig(min::Real, max::Real, dim::Int64)
+function DataConfig(min::Real, max::Real, dim::Integer)
     DataConfig(
         true,               # setup
         repeat([min], dim), # min
@@ -63,24 +96,24 @@ function DataConfig(min::Real, max::Real, dim::Int64)
         dim,                # dim
         dim*2               # dim_comp
     )
-end # DataConfig(min::Real, max::Real, dim::Int64)
+end # DataConfig(min::Real, max::Real, dim::Integer)
 
 """
-    element_min(x::Array, W::Array)
+    element_min(x::RealVector, W::RealVector)
 
 Returns the element-wise minimum between sample x and weight W.
 """
[…]
