+++ src/ARTMAP/SFAM.jl
@@ -15,21 +15,21 @@ Implements a Simple Fuzzy ARTMAP learner's options.
 julia> my_opts = opts_SFAM()
 ```
 """
-@with_kw mutable struct opts_SFAM <: AbstractARTOpts @deftype Float64
+@with_kw mutable struct opts_SFAM <: ARTOpts @deftype RealFP
     # Vigilance parameter: [0, 1]
-    rho = 0.75; @assert rho >= 0 && rho <= 1
+    rho = 0.75; @assert rho >= 0.0 && rho <= 1.0
     # Choice parameter: alpha > 0
-    alpha = 1e-7; @assert alpha > 0
+    alpha = 1e-7; @assert alpha > 0.0
     # Match tracking parameter
-    epsilon = 1e-3; @assert epsilon > 0 && epsilon < 1
+    epsilon = 1e-3; @assert epsilon > 0.0 && epsilon < 1.0
     # Learning parameter: (0, 1]
-    beta = 1; @assert beta > 0 && beta <= 1
+    beta = 1.0; @assert beta > 0.0 && beta <= 1.0
     # Uncommitted node flag
     uncommitted::Bool = true
     # Display flag
     display::Bool = true
-    # Maximum numbers of epochs to train for
-    max_epochs = 1
+    # Maximum number of epochs during training
+    max_epochs::Integer = 1
 end # opts_SFAM()
 
 """
@@ -37,16 +37,16 @@ end # opts_SFAM()
 
 Simple Fuzzy ARTMAP struct.
 """
-mutable struct SFAM <: AbstractART
+mutable struct SFAM <: ART
     opts::opts_SFAM
     config::DataConfig
-    W::Array{Float64, 2}
-    W_old::Array{Float64, 2}
-    labels::Array{Int, 1}
-    y::Array{Int, 1}
-    n_categories::Int
-    epoch::Int
-end # SFAM <: AbstractART
+    W::RealMatrix
+    W_old::RealMatrix
+    labels::IntegerVector
+    y::IntegerVector
+    n_categories::Integer
+    epoch::Integer
+end # SFAM <: ART
 
 """
     SFAM()
@@ -84,17 +84,17 @@ function SFAM(opts::opts_SFAM)
     SFAM(
         opts,                           # opts_SFAM
         DataConfig(),                   # config
-        Array{Float64}(undef, 0,0),     # W
-        Array{Float64}(undef, 0,0),     # W_old
-        Array{Int}(undef, 0),           # labels
-        Array{Int}(undef, 0),           # y
+        Array{RealFP}(undef, 0,0),      # W
+        Array{RealFP}(undef, 0,0),      # W_old
+        Array{Integer}(undef, 0),       # labels
+        Array{Integer}(undef, 0),       # y
         0,                              # n_categories
         0                               # epoch
     )
 end # SFAM(opts::opts_SFAM)
 
 """
-    train!(art::SFAM, x::Array, y::Array ; preprocessed=false)
+    train!(art::SFAM, x::RealArray, y::RealArray ; preprocessed::Bool=false)
 
 Trains a Simple Fuzzy ARTMAP learner in a supervised manner.
 
@@ -108,7 +108,7 @@ SFAM
 julia> train!(art, x, y)
 ```
 """
-function train!(art::SFAM, x::Array, y::Array ; preprocessed=false)
+function train!(art::SFAM, x::RealArray, y::RealArray ; preprocessed::Bool=false)
     # Show a message if display is on
     art.opts.display && @info "Training SFAM"
 
@@ -193,10 +193,10 @@ function train!(art::SFAM, x::Array, y::Array ; preprocessed=false)
         end
         art.W_old = deepcopy(art.W)
     end
[…]
