### 2021-07-01 28af78d
Generalize more types in common, module comments
