+++ src/ARTMAP/ARTMAP.jl
@@ -5,7 +5,7 @@ Description:
     Includes all of the ARTMAP (i.e., explicitly supervised) ART modules definitions.
 """
 
-include("DAM.jl")
-include("FAM.jl")
-include("SFAM.jl")
-include("ARTSCENE.jl")
+include("DAM.jl")       # Default ARTMAP
+include("FAM.jl")       # Fuzzy ARTMAP
+include("SFAM.jl")      # Simplified Fuzzy ARTMAP
+include("ARTSCENE.jl")  # ARTSCENE filters
