+++ src/ART/ART.jl
@@ -5,5 +5,5 @@ Description:
     Includes all of the unsupervised ART modules definitions.
 """
 
-include("DDVFA.jl")
-include("DVFA.jl")
+include("DDVFA.jl")     # DDVFA and GNFA
+include("DVFA.jl")      # DVFA
