### 2021-07-02 c2ba44c
Remove MLJ dependency, redefine performance, bump
