+++ Project.toml
@@ -2,14 +2,13 @@ name = "AdaptiveResonance"
 uuid = "3d72adc0-63d3-4141-bf9b-84450dd0395b"
 authors = ["Sasha Petrenko"]
 description = "A Julia package for Adaptive Resonance Theory (ART) algorithms."
-version = "0.3.0"
+version = "0.3.1"
 
 [deps]
 ClusterValidityIndices = "2fefd308-f647-4698-a2f0-e90cfcbca9ad"
 Distributed = "8ba89e20-285c-5b6f-9357-94700520ee1b"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
-MLJ = "add582a8-e3ab-11e8-2d5e-e98b27df1bc7"
 Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 ProgressBars = "49802e3a-d2f1-5c88-81d8-b72133a6f568"
@@ -18,9 +17,8 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 
 [compat]
 ClusterValidityIndices = "0.1"
-MLJ = "0.14, 0.15, 0.16"
 Parameters = "0.12.1"
-ProgressBars = "0.7, 0.8, 1.0"
+ProgressBars = "0.7, 0.8, 1.0, 1.3"
 julia = "1"
 
 [extras]
