+++ Project.toml
@@ -16,9 +16,9 @@ SharedArrays = "1a1011a3-84de-559e-8e89-a11a2f7dc383"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 
 [compat]
-ClusterValidityIndices = "0.1"
-Parameters = "0.12.1"
-ProgressBars = "0.7, 0.8, 1.0, 1.3"
+ClusterValidityIndices = "0.1, 0.2"
+Parameters = "0.12"
+ProgressBars = "0.7, 0.8, 1"
 julia = "1"
 
 [extras]
