### 2021-07-08 e870445
Remove ClusterValidityIndices.jl requirement
