+++ Project.toml
@@ -2,10 +2,9 @@ name = "AdaptiveResonance"
 uuid = "3d72adc0-63d3-4141-bf9b-84450dd0395b"
 authors = ["Sasha Petrenko"]
 description = "A Julia package for Adaptive Resonance Theory (ART) algorithms."
-version = "0.3.1"
+version = "0.3.2"
 
 [deps]
-ClusterValidityIndices = "2fefd308-f647-4698-a2f0-e90cfcbca9ad"
 Distributed = "8ba89e20-285c-5b6f-9357-94700520ee1b"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
@@ -16,7 +15,6 @@ SharedArrays = "1a1011a3-84de-559e-8e89-a11a2f7dc383"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 
 [compat]
-ClusterValidityIndices = "0.1, 0.2"
 Parameters = "0.12"
 ProgressBars = "0.7, 0.8, 1"
 julia = "1"
