### 2021-07-22 bd0d2b7
Expand upon documentation guide and examples
