### 2021-07-23 #3320 @adi3
@AP6YC no problem, thank you for the list!

@ChrisRackauckas @malmaud @degleris1 @rMassimiliano @aaronpeikert @szkafander - Would any of you be interested in reviewing this paper? We'd greatly appreciate your help!
