### 2021-08-10 #3320 @aaronpeikert
I am currently on vacation. But could review this paper in the week from the 16th August.

A note, though, I am unfamiliar with the review process of JOSS and I am an early career researcher so my experience with reviewing is meagre.
Anyhow, I am quite willing to invest some time to learn how it’s done.
