### 2021-08-10 #3320 @adi3
@aaronpeikert Thank you for the offer! I don't have an issue with waiting till you're back from your vacation. Our review process is pretty well documented and you will be provided with a specific checklist + instructions on how to conduct the review, but in any case, I will be around to guide you through it or have any questions :)

I'll go ahead and assign you as a reviewer on this paper. The actual review will kick off only once we find a second reviewer, upon which you will receive further instructions on how to proceed.
