+++ test/test_utils.jl
@@ -1,4 +1,3 @@
-# using Random
 using DelimitedFiles
 
 """
