### 2021-08-30 #3671 @whedon
[REVIEW]: AdaptiveResonance.jl: A Julia implementation of Adaptive Resonance Theory (ART) algorithms
**Submitting author:** @AP6YC (<a href="http://orcid.org/0000-0003-2442-8901">Sasha Petrenko</a>)
**Repository:** <a href="https://github.com/AP6YC/AdaptiveResonance.jl" target ="_blank">https://github.com/AP6YC/AdaptiveResonance.jl</a>
**Version:** 0.2.2
