### 2021-09-24 490e300
Hotfix DVFA activation norm bug
