### 2021-09-27 be3fc2d
New abstract types, new test methodology
