### 2021-09-27 4ec94ef
Add kwarg constructors for each module
