### 2021-09-30 ad89df9
Up README with quickstart, add tests, bump
