### 2021-10-08 AP6YC/AdaptiveResonance.jl#18 @aaronpeikert
How should users contribute/ask questions?
Hey @AP6YC you really invested a lot of time into making this a usable package! However, when you get some users they will inevitably have some questions.

Maybe you want to point users to where they should ask them (e.g. via issues, StackOverflow or mail).

A Code of Conduct would also be a nice addition:
https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/adding-a-code-of-conduct-to-your-project

I personally like the [Berlin CoC](https://berlincodeofconduct.org).

Related: https://github.com/openjournals/joss-reviews/issues/3671
