### 2021-10-09 AP6YC/AdaptiveResonance.jl#18 @aaronpeikert
GitHub issues might be the best thing for you.
Mentioning them as a place to ask questions and report bugs in Readme and Docs is certainly enough.

Some packages also have this in their startupmessage:

> AdaptiveResonance.jl is beta software. Feel free to ask questions and report bugs at https://github.com/AP6YC/AdaptiveResonance.jl/issues
