### 2021-10-09 AP6YC/AdaptiveResonance.jl#23 @hayesall
[paper] How does AdaptiveResonance.jl compare with other packages?
Here are some questions that could be worth answering in the paper:

- Can you compare outputs between the Matlab and Julia version? (Is the intention of the package to make the Matlab-to-Julia transition fairly painless?)
- Is something appealing about using Julia over Matlab?
- Are there standard references for the Matlab packages mentioned in the README? It could be good to cite them.
