### 2021-10-09 AP6YC/AdaptiveResonance.jl#24 @hayesall
[paper] Statement of need - Target audience
Who is this package for?

- The wider ML community
- the Julia community
- people who are working on ART algorithms who may want benchmarks to compare against?
