### 2021-10-09 AP6YC/AdaptiveResonance.jl#25 @hayesall
[paper] State of the field
The "Tutorial / Guide" documentation page already contains some discussion which could be worth mentioning in the paper.

- For example: supervised vs. unsupervised, batch vs. incremental mode.
- What are the motivations for designing the package with these features?
