### 2021-10-09 AP6YC/AdaptiveResonance.jl#27 @hayesall
[documentation] Minimal working examples
The '*Examples*' page was a fairly helpful entry point to working with this code, but currently it shows *how* rather than *what.* Running the first example produces an error:

```julia
julia> using AdaptiveResonance

julia> train_x, train_y, test_x, test_y = load_some_data()
ERROR: UndefVarError: load_some_data not defined
Stacktrace:
 [1] top-level scope
   @ REPL[6]:1
```

Similar to #26, it would be helpful to setup a motivating problem for each method and show what the expected output might look like.

---

Maybe include something like [`MLDatasets.jl`](https://juliaml.github.io/MLDatasets.jl/stable/) in the documentation? Or maybe re-export some sample datasets from UCI-ML-Repo alongside AdaptiveResonance?
