### 2021-10-09 AP6YC/AdaptiveResonance.jl#28 @hayesall
[documentation] DataConfig parameters / Index
The Index is a little hit-or-miss for me currently. It's useful as a big list of the types and methods, but it's sparse on documentation for what function attributes mean.

---

- At a minimum, I'd like to see functions and types from the [Examples](https://ap6yc.github.io/AdaptiveResonance.jl/dev/man/examples/) to have fairly complete docstrings.
- The [Guide / ART Options](https://ap6yc.github.io/AdaptiveResonance.jl/dev/man/guide/#art_options) are pretty good, some of these would make sense as part of the docstrings for the `opts` objects.
