### 2021-10-09 AP6YC/AdaptiveResonance.jl#18 @hayesall
Some additional feedback on this topic:

The `README / Contributing` section currently reads: "Please raise an issue." It looks like there are some links hidden in the README markdown source such as

```markdown
[contrib-url]: https://juliadocs.github.io/Documenter.jl/dev/contributing/
```

... these might be removed or updated to point to AdaptiveResonance links.

---

Please address each of these in the README (and perhaps on a Developer Notes page #22):

1) **Contribute to the software** --- Do you accept pull requests? Docs, fixes, etc.?
2) **Report issues or problems with the software** --- Issues are good.
3) **Seek support** --- Do you prefer an email or an issue? (e.g. we prefer filing public issues so others can learn or contribute, but questions over email are fine too).
