### 2021-10-12 AP6YC/AdaptiveResonance.jl#24 @aaronpeikert
I think this single issue provides the greatest potential for improving the paper.

If this is for the ML community in Julia the paper should include some problems one can solve with ART.
However, if it is for the ART folks the current paper reads a bit like YAART (**Y**et **A**nother **ART** package) but in my experience, an implementation in Julia has great benefits beyond just being an implementation in another programming language, e.g. julia packages are easier to extend, may benefit from cutting edge numerical computing (GPU/TPU, sparse arrays, symbolic computation etc.) that are not easily available elsewhere, etc.
