### 2021-10-13 7b656a4
Fix typo in dvfa_supervised.jl example script
