+++ docs/src/man/contributing.md
@@ -1,11 +1,60 @@
 # Contributing
 
-To formally contribute to the package, please follow the usual branch pull request procedure:
+## GitFlow
 
-1. Fork the project.
-2. Create your feature branch (git checkout -b my-new-feature).
-3. Commit your changes (git commit -am 'Added some feature').
-4. Push to the branch (git push origin my-new-feature).
-5. Create a new [GitHub pull request](https://github.com/AP6YC/AdaptiveResonance.jl/pulls).
+As of verson `0.3.7`, the `AdaptiveResonance.jl` package follows the [GitFlow](https://nvie.com/posts/a-successful-git-branching-model/) git working model.
+The [original post](https://nvie.com/posts/a-successful-git-branching-model/) by Vincent Driessen outlines this methodology quite well, while [Atlassian](https://www.atlassian.com/git/tutorials/comparing-workflows/gitflow-workflow) has a good tutorial as well.
+In summary:
+
+1. Create a feature branch off of the `develop` branch with the name `feature/<my-feature-name>`.
+2. Commit your changes and push to this feature branch.
+3. When you are satisfied with your changes, initiate a [GitHub pull request](https://github.com/AP6YC/AdaptiveResonance.jl/pulls) (PR) to merge the feature branch with `develop`.
+4. If the unit tests pass, the feature branch will first be merged with develop and then be deleted.
+5. Releases will be periodically initiated from the `develop` branch and versioned onto the `master` branch.
+6. Immediate bug fixes circumvent this process through a `hotfix` branch off of `master`.
+
+## Julia Development
+
+As a Julia package, development follows the usual procedure:
+
+1. Clone the project from GitHub
+2. Switch to or create the branch that you wish work on (see [GitFlow](@ref)).
+3. Start Julia at your development folder.
+4. Instantiate the package (i.e., download and install the package dependencies).
+
+For example, you can get the package and startup Julia with
+
+```sh
+git clone git@github.com:AP6YC/AdaptiveResonance.jl.git
+julia --project=.
+```
+
+!!!note "Note"
+    In Julia, you must activate your project in the current REPL to point to the location/scope of installed packages.
+    The above immediately activates the project when starting up Julia, but you may also separately startup the julia and activate the package with the interactive package manager via the `]` syntax:
+    ```julia
+    julia
+    julia> ]
+    pkg> activate .
+    (AdaptiveResonance) pkg>
+    ```
+
+You may run the package's unit tests after the above setup in Julia with
+
+```julia
+julia> using Pkg
+julia> Pkg.instantiate()
+julia> Pkg.test()
+```
+
+or interactively though the Julia package manager with
+
+```julia
+julia> ]
+(AdaptiveResonance) pkg> instantiate
+(AdaptiveResonance) pkg> test
+```
+
+## Authors
 
 If you simply have suggestions for improvement, Sasha Petrenko (<sap625@mst.edu>) is the current developer and maintainer of the AdaptiveResonance.jl package, so please feel free to reach out with thoughts and questions.
