### 2021-10-13 81507e3
Add development details on contributing page
