+++ docs/src/man/contributing.md
@@ -1,18 +1,18 @@
 # Contributing
 
-## GitFlow
+This page serves as the contribution guide for the `AdaptiveResonance.jl` package.
+From top to bottom, the ways of contributing are:
 
-As of verson `0.3.7`, the `AdaptiveResonance.jl` package follows the [GitFlow](https://nvie.com/posts/a-successful-git-branching-model/) git working model.
-The [original post](https://nvie.com/posts/a-successful-git-branching-model/) by Vincent Driessen outlines this methodology quite well, while [Atlassian](https://www.atlassian.com/git/tutorials/comparing-workflows/gitflow-workflow) has a good tutorial as well.
-In summary:
+- [GitHub Issues:](@ref Issues) how to raise an issue with the project.
+- [Julia Development:](@ref Julia-Development) how to download and interact with the package.
+- [GitFlow:](@ref GitFlow) how to directly contribute code to the package in an organized way on GitHub.
+- [Development Details:](@ref Development-Details) how the internals of the package are currently setup if you would like to directly contribute code.
 
-1. Create a feature branch off of the `develop` branch with the name `feature/<my-feature-name>`.
-2. Commit your changes and push to this feature branch.
-3. When you are satisfied with your changes, initiate a [GitHub pull request](https://github.com/AP6YC/AdaptiveResonance.jl/pulls) (PR) to merge the feature branch with `develop`.
-4. If the unit tests pass, the feature branch will first be merged with develop and then be deleted.
-5. Releases will be periodically initiated from the `develop` branch and versioned onto the `master` branch.
-6. Immediate bug fixes circumvent this process through a `hotfix` branch off of `master`.
+## Issues
 
+The main point of contact is the [GitHub issues](https://github.com/AP6YC/AdaptiveResonance.jl/issues) page for the project.
+This is the easiest way to contribute to the project, as any issue you find or request you have will be addressed there by the authors of the package.
+Depending on the issue, the authors will collaborate with you, and after making changes they will link a [pull request](@ref GitFlow) which addresses your concern or implements your proposed changes.
 ## Julia Development
 
 As a Julia package, development follows the usual procedure:
@@ -29,13 +29,15 @@ git clone git@github.com:AP6YC/AdaptiveResonance.jl.git
 julia --project=.
 ```
 
-!!!note "Note"
+!!! note "Note"
     In Julia, you must activate your project in the current REPL to point to the location/scope of installed packages.
-    The above immediately activates the project when starting up Julia, but you may also separately startup the julia and activate the package with the interactive package manager via the `]` syntax:
+    The above immediately activates the project when starting up Julia, but you may also separately startup the julia and activate the package with the interactive
+    package manager via the `]` syntax:
+
     ```julia
     julia
     julia> ]
-    pkg> activate .
+    (@v1.6) pkg> activate .
     (AdaptiveResonance) pkg>
     ```
 
@@ -53,6 +55,93 @@ or interactively though the Julia package manager with
 julia> ]
 (AdaptiveResonance) pkg> instantiate
 (AdaptiveResonance) pkg> test
+```
+
+## GitFlow
+
+As of verson `0.3.7`, the `AdaptiveResonance.jl` package follows the [GitFlow](https://nvie.com/posts/a-successful-git-branching-model/) git working model.
+The [original post](https://nvie.com/posts/a-successful-git-branching-model/) by Vincent Driessen outlines this methodology quite well, while [Atlassian](https://www.atlassian.com/git/tutorials/comparing-workflows/gitflow-workflow) has a good tutorial as well.
+In summary:
+
+1. Create a feature branch off of the `develop` branch with the name `feature/<my-feature-name>`.
+2. Commit your changes and push to this feature branch.
[…]
