### 2021-10-13 e079b58
Add more information about types and docs
