### 2021-10-13 c18c990
Move and expand contributing in README, delete multiline comments
