### 2021-10-13 AP6YC/AdaptiveResonance.jl#18 @AP6YC
@hayesall Thanks for the catch! PR #37 fixes that boilerplate link in the README, while PR #36 moves and slightly expands upon the contributing section in the README.

Also, PR #35 fleshes out the developer notes in the contributing page. I realized that I didn't address your three bullet points for the README, so I'll address them in one last PR and close this issue.
