### 2021-10-13 ddf6a28
Add longer summary to contributing section in README
