### 2021-10-13 AP6YC/AdaptiveResonance.jl#23 @AP6YC
Thanks for this feedback! I agree and plan on writing up your recommended changes to the paper, which I'll organize on a kanban board in this repo.
