### 2021-10-14 8810d97
Trim table in readme, add logo
