### 2021-10-14 c3b14ba
Add colored logo, banner, and favicon
