### 2021-10-15 9eb49c2
Optimize artwork
