### 2021-10-15 56a6ebc
Smaller favicon
