### 2021-10-15 AP6YC/AdaptiveResonance.jl#26 @AP6YC
Thanks for writing this up! I really like the idea of a small example with figures to show off and bootstrap code for someone unfamiliar with ART. I think that I'd like to take you up on your offer and add it to the docs.
