### 2021-10-15 a5807aa
Add basic example in does with example block
