+++ docs/src/getting-started/whatisart.md
@@ -1,0 +1,5 @@
+# Background
+
+## What is Adaptive Resonance Theory?
+
+TODO
