### 2021-10-15 1e9f47c
Provide some more documentation on all examples
