### 2021-10-15 cb38bbc
Write up ART background and history
