### 2021-10-18 e92611f
Introduce Float alias for largest floating point
