### 2021-10-18 2df7b1b
Add DataConfig convenience function and test
