### 2021-10-18 5ad7491
Replace with Float and signed Int where applicable
