+++ src/ART/DDVFA.jl
@@ -16,7 +16,7 @@ julia> opts_GNFA()
 Initialized GNFA
 ```
 """
-@with_kw mutable struct opts_GNFA <: ARTOpts @deftype RealFP
+@with_kw mutable struct opts_GNFA <: ARTOpts @deftype Float
     # Vigilance parameter: [0, 1]
     rho = 0.6; @assert rho >= 0.0 && rho <= 1.0
     # Choice parameter: alpha > 0
@@ -34,7 +34,7 @@ Initialized GNFA
     # Display flag
     display::Bool = true
     # Maximum number of epochs during training
-    max_epochs::Integer = 1
+    max_epochs::Int = 1
 end # opts_GNFA
 
 """
@@ -56,7 +56,7 @@ mutable struct GNFA <: ART
     config::DataConfig
 
     # Working variables
-    threshold::RealFP
+    threshold::Float
     labels::IntegerVector
     T::RealVector
     M::RealVector
@@ -65,8 +65,8 @@ mutable struct GNFA <: ART
     W::RealMatrix
     W_old::RealMatrix
     n_instance::IntegerVector
-    n_categories::Integer
-    epoch::Integer
+    n_categories::Int
+    epoch::Int
 end # GNFA <: ART
 
 """
@@ -122,12 +122,12 @@ function GNFA(opts::opts_GNFA)
     GNFA(opts,                          # opts
          DataConfig(),                  # config
          0.0,                           # threshold
-         Array{Integer}(undef,0),       # labels
-         Array{RealFP}(undef, 0),       # T
-         Array{RealFP}(undef, 0),       # M
-         Array{RealFP}(undef, 0, 0),    # W
-         Array{RealFP}(undef, 0, 0),    # W_old
-         Array{Integer}(undef, 0),      # n_instance
+         Array{Int}(undef,0),       # labels
+         Array{Float}(undef, 0),       # T
+         Array{Float}(undef, 0),       # M
+         Array{Float}(undef, 0, 0),    # W
+         Array{Float}(undef, 0, 0),    # W_old
+         Array{Int}(undef, 0),      # n_instance
          0,                             # n_categories
          0                              # epoch
     )
@@ -170,7 +170,7 @@ function initialize!(art::GNFA, x::Vector{T} ; y::Integer=0) where {T<:RealFP}
     # IMPORTANT: Assuming that x is a sample, so each entry is a feature
     dim = length(x)
     art.config.dim_comp = dim
-    art.config.dim = Integer(dim/2) # Assumes input is already complement coded
+    art.config.dim = Int(dim/2) # Assumes input is already complement coded
 
     # Initialize the instance and categories counters
     art.n_instance = [1]
@@ -201,7 +201,7 @@ julia> x = load_data()
 julia> train!(my_GNFA, x)
 ```
 """
-function train!(art::GNFA, x::RealArray ; y::IntegerVector = Vector{Integer}())
+function train!(art::GNFA, x::RealArray ; y::IntegerVector = Vector{Int}())
     # Flag for if training in supervised mode
     supervised = !isempty(y)
     # Initialization if weights are empty; fast commit the first sample
@@ -275,7 +275,7 @@ function train!(art::GNFA, x::RealArray ; y::IntegerVector = Vector{Integer}())
         # If we didn't break, deep copy the old weights
         art.W_old = deepcopy(art.W)
     end
-end # train!(art::GNFA, x::RealArray ; y::IntegerVector = Vector{Integer}())
[…]
