+++ .github/workflows/Documentation.yml
@@ -4,6 +4,7 @@ on:
   push:
     branches:
       - master
+      - develop
     tags: '*'
   pull_request:
 
