### 2021-10-19 AP6YC/AdaptiveResonance.jl#41 @markNZed
Warning of method definition
I'm using Julia 1.6.3 and see warnings like:
```
      From worker 2:	WARNING: Method definition (::Type{AdaptiveResonance.FAM})() in module AdaptiveResonance at /Users/XXX/AdaptiveResonance.jl/src/ARTMAP/FAM.jl:82 overwritten at /Users/XXX/AdaptiveResonance.jl/src/ARTMAP/FAM.jl:64.
      From worker 2:	  ** incremental compilation may be fatally broken for this module **
```
It does not seem to break things but it would be nice if we did not have these warnings.
