### 2021-10-19 AP6YC/AdaptiveResonance.jl#42 @hayesall
[bug?] DDVFA mismatch_flag should probably be false in supervised classification
(*I might be misinterpreting something, but this seemed like something that should not occur in the supervised case.*)

In rare cases, running `classify` with supervised DDVFA may return a vector containing `-1` that does not correspond to one of the labels in the original data space.

![Plot of supervised DDVFA with two Gaussians. There are three labels: 0, 1, -1; but -1 was not part of the original feature space](https://user-images.githubusercontent.com/11916674/137928746-46696041-46d6-4736-a4b9-4935bec6d38d.png)

It looks like this is caused by an interaction between `mismatch_flag` defaulted to `true`, and the similarity/`art.threshold` value:

https://github.com/AP6YC/AdaptiveResonance.jl/blob/f1227ea692643796ae3d38eb7267b0536f928688/src/ART/DDVFA.jl#L818-L822

## Reproducing

This only seems to happen ~30% of the time depending on the split of the data:

```julia
julia> DataStructures.counter(y_hat_test)
Accumulator{Int64, Int64} with 3 entries:
  0  => 304
  -1 => 2
  1  => 294
```

But here's the code I used to create the figures:

```julia
using AdaptiveResonance
using Distributions, Random
using MLDataUtils
using Plots

rng = MersenneTwister(1234)
dist1 = MvNormal([0.0, 6.0], [1.0 0.0; 0.0 1.0])
dist2 = MvNormal([4.5, 6.0], [2.0 -1.5; -1.5 2.0])

N_POINTS = 1000

X = hcat(rand(rng, dist1, N_POINTS), rand(rng, dist2, N_POINTS))
y = vcat(ones(Int64, N_POINTS), zeros(Int64, N_POINTS))

p1 = scatter(X[1,:], X[2,:], group=y, title="Original Data")

(X_train, y_train), (X_test, y_test) = stratifiedobs((X, y))

art = DDVFA()
train!(art, X_train)
y_hat_test = AdaptiveResonance.classify(art, X_test)
p2 = scatter(X_test[1,:], X_test[2,:], group=y_hat_test, title="Unsupervised DDVFA")

art = DDVFA()
train!(art, X_train, y=y_train)
y_hat_test = AdaptiveResonance.classify(art, X_test)
p3 = scatter(X_test[1,:], X_test[2,:], group=y_hat_test, title="Supervised DDVFA", xlabel="Performance: " * string(round(performance(y_hat_test, y_test); digits=3)))

plt = plot(p1, p2, p3, layout=(1, 3), legend = false, xtickfontsize=6, xguidefontsize=8, titlefont=font(8))
```
