### 2021-10-20 5ed4681
Move GNFA to separate file as its own module
