+++ src/AdaptiveResonance.jl
@@ -46,8 +46,8 @@ export
     get_n_samples,              # Get the number of samples (1-D interpreted as one sample)
 
     # ART (unsupervised)
-    DDVFA, opts_DDVFA,
     GNFA, opts_GNFA,
+    DDVFA, opts_DDVFA,
     DVFA, opts_DVFA,
 
     # ARTMAP (supervised)
