+++ src/ART/DDVFA.jl
@@ -6,387 +6,6 @@ Description:
 """
 
 """
-    opts_GNFA()
-
-Gamma-Normalized Fuzzy ART options struct.
-
-# Examples
-```julia-repl
-julia> opts_GNFA()
-Initialized GNFA
-```
-"""
-@with_kw mutable struct opts_GNFA <: ARTOpts @deftype Float
-    # Vigilance parameter: [0, 1]
-    rho = 0.6; @assert rho >= 0.0 && rho <= 1.0
-    # Choice parameter: alpha > 0
-    alpha = 1e-3; @assert alpha > 0.0
-    # Learning parameter: (0, 1]
-    beta = 1.0; @assert beta > 0.0 && beta <= 1.0
-    # "Pseudo" kernel width: gamma >= 1
-    gamma = 3.0; @assert gamma >= 1.0
-    # gamma = 784; @assert gamma >= 1
-    # "Reference" gamma for normalization: 0 <= gamma_ref < gamma
-    gamma_ref = 1.0; @assert 0.0 <= gamma_ref && gamma_ref < gamma
-    # Similarity method (activation and match):
-    #   'single', 'average', 'complete', 'median', 'weighted', or 'centroid'
-    method::String = "single"
-    # Display flag
-    display::Bool = true
-    # Maximum number of epochs during training
-    max_epochs::Int = 1
-end # opts_GNFA
-
-"""
-    GNFA <: ART
-
-Gamma-Normalized Fuzzy ART learner struct
-
-# Examples
-```julia-repl
-julia> GNFA()
-GNFA
-    opts: opts_GNFA
-    ...
-```
-"""
-mutable struct GNFA <: ART
-    # Assign numerical parameters from options
-    opts::opts_GNFA
-    config::DataConfig
-
-    # Working variables
-    threshold::Float
-    labels::IntegerVector
-    T::RealVector
-    M::RealVector
-
-    # "Private" working variables
-    W::RealMatrix
-    W_old::RealMatrix
-    n_instance::IntegerVector
-    n_categories::Int
-    epoch::Int
-end # GNFA <: ART
-
-"""
-    GNFA()
-
-Implements a Gamma-Normalized Fuzzy ART learner.
-
-# Examples
-```julia-repl
-julia> GNFA()
-GNFA
-    opts: opts_GNFA
-    ...
-```
-"""
-function GNFA()
-    opts = opts_GNFA()
-    GNFA(opts)
-end # GNFA()
-
-"""
-    GNFA(;kwargs...)
-
-Implements a Gamma-Normalized Fuzzy ART learner with keyword arguments.
-
-# Examples
-```julia-repl
-julia> GNFA(rho=0.7)
-GNFA
-    opts: opts_GNFA
-    ...
-```
-"""
-function GNFA(;kwargs...)
-    opts = opts_GNFA(;kwargs...)
-    GNFA(opts)
-end # GNFA(;kwargs...)
-
-"""
-    GNFA(opts::opts_GNFA)
-
-Implements a Gamma-Normalized Fuzzy ART learner with specified options.
-
-# Examples
-```julia-repl
-julia> GNFA(opts)
-GNFA
-    opts: opts_GNFA
-    ...
-```
-"""
-function GNFA(opts::opts_GNFA)
-    GNFA(opts,                          # opts
-         DataConfig(),                  # config
-         0.0,                           # threshold
-         Array{Int}(undef,0),       # labels
-         Array{Float}(undef, 0),       # T
-         Array{Float}(undef, 0),       # M
-         Array{Float}(undef, 0, 0),    # W
-         Array{Float}(undef, 0, 0),    # W_old
-         Array{Int}(undef, 0),      # n_instance
-         0,                             # n_categories
[…]
