+++ src/ART/ART.jl
@@ -5,5 +5,6 @@ Description:
     Includes all of the unsupervised ART modules definitions.
 """
 
+include("GNFA.jl")      # GNFA
 include("DDVFA.jl")     # DDVFA and GNFA
 include("DVFA.jl")      # DVFA
