### 2021-10-20 a82b812
Copy GNFA to FuzzyART as template
