+++ src/ART/FuzzyART.jl
@@ -1,22 +1,22 @@
 """
-    GNFA.jl
+    FuzzyART.jl
 
 Description:
     Includes all of the structures and logic for running a Gamma-Normalized Fuzzy ART module.
 """
 
 """
-    opts_GNFA()
+    opts_FuzzyART()
 
 Gamma-Normalized Fuzzy ART options struct.
 
 # Examples
 ```julia-repl
-julia> opts_GNFA()
-Initialized GNFA
+julia> opts_FuzzyART()
+Initialized FuzzyART
 ```
 """
-@with_kw mutable struct opts_GNFA <: ARTOpts @deftype Float
+@with_kw mutable struct opts_FuzzyART <: ARTOpts @deftype Float
     # Vigilance parameter: [0, 1]
     rho = 0.6; @assert rho >= 0.0 && rho <= 1.0
     # Choice parameter: alpha > 0
@@ -35,24 +35,24 @@ Initialized GNFA
     display::Bool = true
     # Maximum number of epochs during training
     max_epochs::Int = 1
-end # opts_GNFA
+end # opts_FuzzyART
 
 """
-    GNFA <: ART
+    FuzzyART <: ART
 
 Gamma-Normalized Fuzzy ART learner struct
 
 # Examples
 ```julia-repl
-julia> GNFA()
-GNFA
-    opts: opts_GNFA
+julia> FuzzyART()
+FuzzyART
+    opts: opts_FuzzyART
     ...
 ```
 """
-mutable struct GNFA <: ART
+mutable struct FuzzyART <: ART
     # Assign numerical parameters from options
-    opts::opts_GNFA
+    opts::opts_FuzzyART
     config::DataConfig
 
     # Working variables
@@ -67,59 +67,59 @@ mutable struct GNFA <: ART
     n_instance::IntegerVector
     n_categories::Int
     epoch::Int
-end # GNFA <: ART
+end # FuzzyART <: ART
 
 """
-    GNFA()
+    FuzzyART()
 
 Implements a Gamma-Normalized Fuzzy ART learner.
 
 # Examples
 ```julia-repl
-julia> GNFA()
-GNFA
-    opts: opts_GNFA
+julia> FuzzyART()
+FuzzyART
+    opts: opts_FuzzyART
     ...
 ```
 """
-function GNFA()
-    opts = opts_GNFA()
-    GNFA(opts)
-end # GNFA()
+function FuzzyART()
+    opts = opts_FuzzyART()
+    FuzzyART(opts)
+end # FuzzyART()
 
 """
-    GNFA(;kwargs...)
+    FuzzyART(;kwargs...)
 
 Implements a Gamma-Normalized Fuzzy ART learner with keyword arguments.
 
 # Examples
 ```julia-repl
-julia> GNFA(rho=0.7)
-GNFA
-    opts: opts_GNFA
+julia> FuzzyART(rho=0.7)
+FuzzyART
+    opts: opts_FuzzyART
     ...
 ```
 """
-function GNFA(;kwargs...)
-    opts = opts_GNFA(;kwargs...)
-    GNFA(opts)
-end # GNFA(;kwargs...)
+function FuzzyART(;kwargs...)
+    opts = opts_FuzzyART(;kwargs...)
+    FuzzyART(opts)
+end # FuzzyART(;kwargs...)
 
 """
-    GNFA(opts::opts_GNFA)
+    FuzzyART(opts::opts_FuzzyART)
 
 Implements a Gamma-Normalized Fuzzy ART learner with specified options.
 
 # Examples
 ```julia-repl
-julia> GNFA(opts)
-GNFA
-    opts: opts_GNFA
+julia> FuzzyART(opts)
+FuzzyART
+    opts: opts_FuzzyART
     ...
 ```
 """
[…]
