### 2021-10-20 750fc09
FuzzyART init sample is RealVector
