### 2021-10-20 159f6ed
Add get_W to DDVFA, remove deepcopy stop condition
