### 2021-10-20 2dbe170
Remove W_old from FuzzyART and GNFA
