+++ src/ART/GNFA.jl
@@ -63,7 +63,6 @@ mutable struct GNFA <: ART
 
     # "Private" working variables
     W::RealMatrix
-    W_old::RealMatrix
     n_instance::IntegerVector
     n_categories::Int
     epoch::Int
@@ -126,7 +125,6 @@ function GNFA(opts::opts_GNFA)
         Array{Float}(undef, 0),         # T
         Array{Float}(undef, 0),         # M
         Array{Float}(undef, 0, 0),      # W
-        Array{Float}(undef, 0, 0),      # W_old
         Array{Int}(undef, 0),           # n_instance
         0,                              # n_categories
         0                               # epoch
@@ -213,8 +211,6 @@ function train!(art::GNFA, x::RealArray ; y::IntegerVector = Vector{Int}())
         skip_first = false
     end
 
-    art.W_old = deepcopy(art.W)
-
     # Learning
     art.epoch = 0
     while true
@@ -271,8 +267,6 @@ function train!(art::GNFA, x::RealArray ; y::IntegerVector = Vector{Int}())
         if stopping_conditions(art)
             break
         end
-        # If we didn't break, deep copy the old weights
-        art.W_old = deepcopy(art.W)
     end
 end # train!(art::GNFA, x::RealArray ; y::IntegerVector = Vector{Int}())
 
@@ -382,5 +376,5 @@ end # learn!(art::GNFA, x::RealVector, index::Integer)
 Stopping conditions for a GNFA module.
 """
 function stopping_conditions(art::GNFA)
-    return isequal(art.W, art.W_old) || art.epoch >= art.opts.max_epochs
+    return art.epoch >= art.opts.max_epochs
 end # stopping_conditions(art::GNFA)
