### 2021-10-20 81b4562
Refactor initialization logic, seq and batch
