### 2021-10-21 a7d2988
Overhaul GNFA and common
