+++ src/ART/GNFA.jl
@@ -27,7 +27,7 @@ Initialized GNFA
     gamma = 3.0; @assert gamma >= 1.0
     # gamma = 784; @assert gamma >= 1
     # "Reference" gamma for normalization: 0 <= gamma_ref < gamma
-    gamma_ref = 1.0; @assert 0.0 <= gamma_ref && gamma_ref < gamma
+    gamma_ref = 1.0; @assert 0.0 <= gamma_ref && gamma_ref <= gamma
     # Similarity method (activation and match):
     #   'single', 'average', 'complete', 'median', 'weighted', or 'centroid'
     method::String = "single"
@@ -168,7 +168,7 @@ function initialize!(art::GNFA, x::Vector{T} ; y::Integer=0) where {T<:RealFP}
     art.W = Array{T}(undef, art.config.dim_comp, 1)
     # Assign the contents, valid this way for 1-D or 2-D arrays
     art.W[:, 1] = x
-    label = iszero(y) ? y : 1
+    label = !iszero(y) ? y : 1
     push!(art.labels, label)
 end # initialize!(art::GNFA, x::Vector{T} ; y::Integer=0) where {T<:RealFP}
 
@@ -185,8 +185,7 @@ function train!(art::GNFA, x::RealVector ; y::Integer = 0, preprocessed::Bool=fa
     # Initialization if weights are empty; fast commit the first sample
     if isempty(art.W)
         label = supervised ? y : 1
-        push!(art.labels, label)
-        initialize!(art, x)
+        initialize!(art, x, y=label)
         return
     end
 
@@ -196,17 +195,27 @@ function train!(art::GNFA, x::RealVector ; y::Integer = 0, preprocessed::Bool=fa
     index = sortperm(art.T, rev=true)
     # Initialize mismatch as true
     mismatch_flag = true
+
+    # If we have a new supervised category, create a new category
+    if supervised && !(y in art.labels)
+        create_category(art, x, y)
+        return y
+    end
     # Loop over all categories
     for j = 1:art.n_categories
         # Best matching unit
         bmu = index[j]
         # Vigilance check - pass
         if art.M[bmu] >= art.threshold
+            # If supervised and the label differed, force mismatch
+            if supervised && (art.labels[bmu] != y)
+                break
+            end
             # Learn the sample
             learn!(art, x, bmu)
             # Update sample labels
-            label = supervised ? y : bmu
-            push!(art.labels, label)
+            # label = supervised ? y : bmu
+            # push!(art.labels, label)
             # No mismatch
             mismatch_flag = false
             break
@@ -214,19 +223,28 @@ function train!(art::GNFA, x::RealVector ; y::Integer = 0, preprocessed::Bool=fa
     end
     # If there was no resonant category, make a new one
     if mismatch_flag
-        # Increment the number of categories
-        art.n_categories += 1
-        # Fast commit
-        art.W = hcat(art.W, x)
-        # Increment number of samples associated with new category
-        push!(art.n_instance, 1)
-        # Update sample labels
-        label = supervised ? y : art.n_categories
-        push!(art.labels, label)
+        # Get the correct label for the new category
+        label = supervised ? y : art.n_categories + 1
+        # Create a new category
+        create_category(art, x, label)
     end
 
-    return
+    return y
 end # train!(art::GNFA, x::RealVector ; y::Integer = 0, preprocessed::Bool=false)
+
+"""
+    create_category(art::GNFA, x::RealVector, y::Integer)
+"""
+function create_category(art::GNFA, x::RealVector, y::Integer)
+    # Increment the number of categories
[…]
