### 2021-10-21 3c2b712
Add convenience functions for DDVFA, trim
