### 2021-10-21 b4c7876
Add get_bmu to ARTMAPs, remove redundant tests
