+++ test/test_ddvfa.jl
@@ -1,190 +1,0 @@
-"""
-    tt_ddvfa(opts::opts_DDVFA, train_x::Array)
-
-Trains and tests (tt) a DDVFA module on unlabeled data train_x.
-"""
-function tt_ddvfa(opts::opts_DDVFA, train_x::Array)
-    # Create the ART module, train, and classify
-    art = DDVFA(opts)
-    train!(art, train_x)
-    y_hat = classify(art, train_x)
-    # perf = performance(y_hat, data.test_y)
-
-    # Total number of categories
-    total_vec = [art.F2[i].n_categories for i = 1:art.n_categories]
-    total_cat = sum(total_vec)
-    @info "Categories: $(art.n_categories)"
-    @info "Weights: $total_cat"
-
-    return art
-end # tt_ddvfa(opts::opts_DDVFA, train_x::Array)
-
-@testset "DDVFA Sequential" begin
-    @info "------- DDVFA Sequential -------"
-
-    # Initialize the ART module
-    art = DDVFA()
-    # Turn off display for sequential training/testing
-    art.opts.display = false
-    # Set up the data manually because the module can't infer from single samples
-    data_setup!(art.config, data.train_x)
-
-    # Get the dimension and size of the data
-    dim, n_samples = get_data_shape(data.train_x)
-    y_hat_train = zeros(Int64, n_samples)
-    dim_test, n_samples_test = get_data_shape(data.test_x)
-    y_hat = zeros(Int64, n_samples_test)
-    y_hat_bmu = zeros(Int64, n_samples_test)
-
-    # Iterate over all examples sequentially
-    for i = 1:n_samples
-        y_hat_train[i] = train!(art, data.train_x[:, i], y=[data.train_y[i]])
-    end
-
-    # Iterate over all test samples sequentially
-    for i = 1:n_samples_test
-        y_hat[i] = classify(art, data.test_x[:, i])
-        y_hat_bmu[i] = classify(art, data.test_x[:, i], get_bmu=true)
-    end
-
-    # Calculate performance
-    perf_train = performance(y_hat_train, data.train_y)
-    perf_test = performance(y_hat, data.test_y)
-    perf_test_bmu = performance(y_hat_bmu, data.test_y)
-
-    # Test the permance above a baseline number
-    perf_baseline = 0.8
-    @test perf_train >= perf_baseline
-    @test perf_test >= perf_baseline
-    @test perf_test_bmu >= perf_baseline
-
-    @info "DDVFA Training Perf: $perf_train"
-    @info "DDVFA Testing Perf: $perf_test"
-    @info "DDVFA Testing BMU Perf: $perf_test_bmu"
-end
-
-@testset "DDVFA Supervised" begin
-    @info "------- DDVFA Supervised -------"
-
-    # Train and classify
-    art = DDVFA()
-    y_hat_train = train!(art, data.train_x, y=data.train_y)
-    y_hat = classify(art, data.test_x)
-    y_hat_bmu = classify(art, data.test_x, get_bmu=true)
-
-    # Calculate performance
-    perf_train = performance(y_hat_train, data.train_y)
-    perf_test = performance(y_hat, data.test_y)
-    perf_test_bmu = performance(y_hat_bmu, data.test_y)
-
-    # Test the performances with a baseline number
-    perf_baseline = 0.8
-    @test perf_train >= perf_baseline
-    @test perf_test >= perf_baseline
-    @test perf_test_bmu >= perf_baseline
-
-    # Log the results
-    @info "DDVFA Training Perf: $perf_train"
-    @info "DDVFA Testing Perf: $perf_test"
-    @info "DDVFA Testing BMU Perf: $perf_test_bmu"
[…]
