+++ src/ARTMAP/DAM.jl
@@ -229,7 +229,7 @@ julia> train!(art, x, y)
 julia> classify(art, x_test)
 ```
 """
-function classify(art::DAM, x::RealMatrix ; preprocessed::Bool=false)
+function classify(art::DAM, x::RealMatrix ; preprocessed::Bool=false, get_bmu::Bool=false)
     # Show a message if display is on
     art.opts.display && @info "Testing DAM"
 
@@ -274,7 +274,13 @@ function classify(art::DAM, x::RealMatrix ; preprocessed::Bool=false)
         if mismatch_flag
             # Create new weight vector
             @debug "Mismatch"
-            y_hat[ix] = -1
+            # If we mismatched and want the best matching unit
+            if get_bmu
+                y_hat = art.labels[index[1]]
+            # Otherwise, report the mismatch label -1
+            else
+                y_hat[ix] = -1
+            end
         end
     end
     return y_hat
