### 2021-10-21 b334eff
Remove redundant train/test method
