### 2021-10-21 d6bb05f
Add threshold_normalization option to DDVFA, GNFA
