+++ test/test_sets.jl
@@ -66,9 +66,20 @@ end # @testset "AdaptiveResonance.jl"
     # All common ART options
     art_opts = [
         (display = true,),
-        (display = false,),
+        # (display = false,),
     ]
-    n_art_opts = length(art_opts)
+
+    # Specific ART options
+    art_specifics = Dict(
+        DDVFA => [
+            (threshold_normalization=true,),
+            (threshold_normalization=false,),
+        ],
+        GNFA => [
+            (threshold_normalization=true,),
+            (threshold_normalization=false,),
+        ],
+    )
 
     # All test option permutations
     test_opts = [
@@ -85,15 +96,21 @@ end # @testset "AdaptiveResonance.jl"
     for ix = 1:n_arts
         # Iterate over all test options
         for jx = 1:n_test_opts
-            for kx = 1:n_art_opts
+            # If we are testing a module with different options, merge
+            if haskey(art_specifics, arts[ix])
+                local_art_opts = vcat(art_opts, art_specifics[arts[ix]])
+            else
+                local_art_opts = art_opts
+            end
+            # Iterate over all options
+            for kx = 1:length(local_art_opts)
                 # Only do the unsupervised method if we have an ART module (not ARTMAP)
                 if arts[ix] isa ART
                     # Unsupervised
-                    train_test_art(arts[ix](;art_opts[kx]...), data; test_opts=test_opts[jx])
+                    train_test_art(arts[ix](;local_art_opts[kx]...), data; test_opts=test_opts[jx])
                 end
-
                 # Supervised
-                @test train_test_art(arts[ix](;art_opts[kx]...), data; supervised=true, test_opts=test_opts[jx]) >= perf_baseline
+                @test train_test_art(arts[ix](;local_art_opts[kx]...), data; supervised=true, test_opts=test_opts[jx]) >= perf_baseline
             end
         end
     end
