### 2021-10-22 582ea16
Remove FuzzyART file to refactor GNFA
