+++ src/ART/FuzzyART.jl
@@ -1,381 +1,0 @@
-"""
-    FuzzyART.jl
-
-Description:
-    Includes all of the structures and logic for running a Gamma-Normalized Fuzzy ART module.
-"""
-
-"""
-    opts_FuzzyART()
-
-Gamma-Normalized Fuzzy ART options struct.
-
-# Examples
-```julia-repl
-julia> opts_FuzzyART()
-Initialized FuzzyART
-```
-"""
-@with_kw mutable struct opts_FuzzyART <: ARTOpts @deftype Float
-    # Vigilance parameter: [0, 1]
-    rho = 0.6; @assert rho >= 0.0 && rho <= 1.0
-    # Choice parameter: alpha > 0
-    alpha = 1e-3; @assert alpha > 0.0
-    # Learning parameter: (0, 1]
-    beta = 1.0; @assert beta > 0.0 && beta <= 1.0
-    # "Pseudo" kernel width: gamma >= 1
-    gamma = 3.0; @assert gamma >= 1.0
-    # gamma = 784; @assert gamma >= 1
-    # "Reference" gamma for normalization: 0 <= gamma_ref < gamma
-    gamma_ref = 1.0; @assert 0.0 <= gamma_ref && gamma_ref < gamma
-    # Similarity method (activation and match):
-    #   'single', 'average', 'complete', 'median', 'weighted', or 'centroid'
-    method::String = "single"
-    # Display flag
-    display::Bool = true
-    # Maximum number of epochs during training
-    max_epochs::Int = 1
-end # opts_FuzzyART
-
-"""
-    FuzzyART <: ART
-
-Gamma-Normalized Fuzzy ART learner struct
-
-# Examples
-```julia-repl
-julia> FuzzyART()
-FuzzyART
-    opts: opts_FuzzyART
-    ...
-```
-"""
-mutable struct FuzzyART <: ART
-    # Assign numerical parameters from options
-    opts::opts_FuzzyART
-    config::DataConfig
-
-    # Working variables
-    threshold::Float
-    labels::IntegerVector
-    T::RealVector
-    M::RealVector
-
-    # "Private" working variables
-    W::RealMatrix
-    n_instance::IntegerVector
-    n_categories::Int
-    epoch::Int
-end # FuzzyART <: ART
-
-"""
-    FuzzyART()
-
-Implements a Gamma-Normalized Fuzzy ART learner.
-
-# Examples
-```julia-repl
-julia> FuzzyART()
-FuzzyART
-    opts: opts_FuzzyART
-    ...
-```
-"""
-function FuzzyART()
-    opts = opts_FuzzyART()
-    FuzzyART(opts)
-end # FuzzyART()
-
-"""
-    FuzzyART(;kwargs...)
-
-Implements a Gamma-Normalized Fuzzy ART learner with keyword arguments.
-
-# Examples
-```julia-repl
-julia> FuzzyART(rho=0.7)
-FuzzyART
-    opts: opts_FuzzyART
-    ...
-```
-"""
-function FuzzyART(;kwargs...)
-    opts = opts_FuzzyART(;kwargs...)
-    FuzzyART(opts)
-end # FuzzyART(;kwargs...)
-
-"""
-    FuzzyART(opts::opts_FuzzyART)
-
-Implements a Gamma-Normalized Fuzzy ART learner with specified options.
-
-# Examples
-```julia-repl
-julia> FuzzyART(opts)
-FuzzyART
-    opts: opts_FuzzyART
-    ...
-```
-"""
-function FuzzyART(opts::opts_FuzzyART)
-    FuzzyART(opts,                      # opts
-        DataConfig(),                   # config
-        0.0,                            # threshold
-        Array{Int}(undef,0),            # labels
-        Array{Float}(undef, 0),         # T
[…]
