+++ src/ART/ART.jl
@@ -5,7 +5,6 @@ Description:
     Includes all of the unsupervised ART modules definitions.
 """
 
-include("FuzzyART.jl")  # FuzzyART
 include("GNFA.jl")      # GNFA
 include("DDVFA.jl")     # DDVFA
 include("DVFA.jl")      # DVFA
