### 2021-10-22 0fbed6d
Refactor GNFA to FuzzyART
