+++ src/ART/ART.jl
@@ -5,6 +5,6 @@ Description:
     Includes all of the unsupervised ART modules definitions.
 """
 
-include("GNFA.jl")      # GNFA
+include("FuzzyART.jl")  # FuzzyART
 include("DDVFA.jl")     # DDVFA
 include("DVFA.jl")      # DVFA
