### 2021-10-22 f5ee58f
Add modules section to docs and README
