### 2021-10-22 96d0cb4
Rename threshold_normalization to gamma_normalization
