### 2021-10-25 b530a68
Refactor DDVFA train!, clean FuzzyART
