+++ src/ART/DDVFA.jl
@@ -21,8 +21,7 @@ julia> my_opts = opts_DDVFA()
 """
 @with_kw mutable struct opts_DDVFA <: ARTOpts @deftype Float
     # Lower-bound vigilance parameter: [0, 1]
-    rho_lb = 0.80; @assert rho_lb >= 0.0 && rho_lb <= 1.0
-    # rho = rho_lb
+    rho_lb = 0.7; @assert rho_lb >= 0.0 && rho_lb <= 1.0
     # Upper bound vigilance parameter: [0, 1]
     rho_ub = 0.85; @assert rho_ub >= 0.0 && rho_ub <= 1.0
     # Choice parameter: alpha > 0
@@ -169,66 +168,112 @@ end # DDVFA(opts::opts_DDVFA)
 Sets the vigilance threshold of the DDVFA module as a function of several flags and hyperparameters.
 """
 function set_threshold!(art::DDVFA)
+    # Gamma match normalization
     if art.opts.gamma_normalization
+        # Set the learning threshold as a function of the data dimension
         art.threshold = art.opts.rho_lb*(art.config.dim^art.opts.gamma_ref)
     else
+        # Set the learning threshold as simply the vigilance parameter
         art.threshold = art.opts.rho_lb
     end
 end # set_threshold!(art::DDVFA)
 
 """
-    train!(art::DDVFA, x::RealArray ; y::IntegerVector=Vector{Int}(), preprocessed::Bool=false)
+    train!(art::DDVFA, x::RealVector ; y::Integer=0, preprocessed::Bool=false)
+"""
+function train!(art::DDVFA, x::RealVector ; y::Integer=0, preprocessed::Bool=false)
+    # Flag for if training in supervised mode
+    supervised = !iszero(y)
+
+    # Run the sequential initialization procedure
+    sample = init_train!(x, art, preprocessed)
+
+    # Initialization
+    if isempty(art.F2)
+        # Set the first label as either 1 or the first provided label
+        y_hat = supervised ? y : 1
+        # Create a new category
+        create_category(art, sample, y_hat)
+        return y_hat
+    end
+
+    # # If we have a new supervised category, create a new category
+    # if supervised && !(y in art.labels)
+    #     create_category(art, sample, y)
+    #     return y
+    # end
+
+    # Default to mismatch
+    mismatch_flag = true
+
+    # Compute the activation for all categories
+    T = zeros(art.n_categories)
+    for jx = 1:art.n_categories
+        activation_match!(art.F2[jx], sample)
+        T[jx] = similarity(art.opts.method, art.F2[jx], "T", sample, art.opts.gamma_ref)
+    end
+
+    # Compute the match for each category in the order of greatest activation
+    index = sortperm(T, rev=true)
+    for jx = 1:art.n_categories
+        bmu = index[jx]
+        M = similarity(art.opts.method, art.F2[bmu], "M", sample, art.opts.gamma_ref)
+        # If we got a match, then learn (update the category)
+        if M >= art.threshold
+            # Update the stored match and activation values
+            art.M = M
+            art.T = T[bmu]
+            # If supervised and the label differs, trigger mismatch
+            if supervised && (art.labels[bmu] != y)
+                break
+            end
+            # Update the weights with the sample
+            train!(art.F2[bmu], sample, preprocessed=true)
+            # Save the output label for the sample
+            y_hat = art.labels[bmu]
+            # No mismatch
+            mismatch_flag = false
+            break
+        end
+    end
+
+    # If we triggered a mismatch
+    if mismatch_flag
+        # Update the stored match and activation values
+        bmu = index[1]
[…]
