### 2021-10-25 ef6f799
Refactor DDVFA classify
