### 2021-10-25 2ad14c3
Refactor DVFA, generalize ART batch methods
