+++ src/ART/DVFA.jl
@@ -60,11 +60,12 @@ mutable struct DVFA <: ART
     config::DataConfig
 
     # Working variables
+    threshold_ub::Float
+    threshold_lb::Float
     labels::IntegerVector
     W::RealMatrix
     T::RealVector
     M::RealVector
-    W_old::RealMatrix
     map::IntegerVector
     n_categories::Int
     n_clusters::Int
@@ -125,11 +126,12 @@ function DVFA(opts::opts_DVFA)
     DVFA(
         opts,                           # opts
         DataConfig(),                   # config
+        0.0,                            # threshold_ub
+        0.0,                            # threshold_lb
         Array{Int}(undef, 0),           # labels
         Array{Float}(undef, 0, 0),      # W
         Array{Float}(undef, 0),         # M
         Array{Float}(undef, 0),         # T
-        Array{Float}(undef, 0, 0),      # W_old
         Array{Int}(undef, 0),           # map
         0,                              # n_categories
         0,                              # n_clusters
@@ -137,8 +139,14 @@ function DVFA(opts::opts_DVFA)
     )
 end # DDVFA(opts::opts_DDVFA)
 
+function set_threshold!(art::DVFA)
+    # DVFA thresholds
+    art.threshold_ub = art.opts.rho_ub * art.config.dim
+    art.threshold_lb = art.opts.rho_lb * art.config.dim
+end
+
 """
-    train!(art::DVFA, x::RealArray ; y::IntegerVector = Vector{Int}(), preprocessed::Bool=false)
+    train!(art::DVFA, x::RealMatrix ; y::IntegerVector=Vector{Int}(), preprocessed::Bool=false)
 
 Train the DVFA module on x with optional custom category labels y.
 
@@ -147,144 +155,99 @@ Train the DVFA module on x with optional custom category labels y.
 - `x::RealArray`: the data to train on, interpreted as a single sample if x is a vector.
 - `y::IntegerVector=[]`: optional custom labels to assign to the categories. If empty, ordinary incremental labels are prescribed.
 """
-function train!(art::DVFA, x::RealArray ; y::IntegerVector = Vector{Int}(), preprocessed::Bool=false)
-    # Show a message if display is on
-    art.opts.display && @info "Training DVFA"
+function train!(art::DVFA, x::RealVector ; y::Integer=0, preprocessed::Bool=false)
 
-    # Simple supervised flag
-    supervised = !isempty(y)
+    # Flag for if training in supervised mode
+    supervised = !iszero(y)
 
-    # Data information and setup
-    n_samples = get_n_samples(x)
-
-    # Set up the data config if training for the first time
-    !art.config.setup && data_setup!(art.config, x)
-
-    # If the data is not preprocessed, then complement code it
-    if !preprocessed
-        x = complement_code(x, config=art.config)
-    end
-
-    if n_samples == 1
-        y_hat = zero(Int)
-    else
-        y_hat = zeros(Int, n_samples)
-    end
+    # Run the sequential initialization procedure
+    sample = init_train!(x, art, preprocessed)
 
     # Initialization
     if isempty(art.W)
+        # Set the threshold
+        set_threshold!(art)
         # Set the first label as either 1 or the first provided label
-        local_label = supervised ? y[1] : 1
-        # Add the local label to the output vector
-        if n_samples == 1
-            y_hat = local_label
-        else
-            y_hat[1] = local_label
-        end
+        y_hat = supervised ? y : 1
         # Create a new category and cluster
         art.W = ones(art.config.dim_comp, 1)
         art.n_categories = 1
         art.n_clusters = 1
-        push!(art.labels, local_label)
-        # Skip the first training entry
-        skip_first = true
-    else
[…]
