### 2021-10-25 8a32699
Clean docstrings, move common ART methods
