+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"man/modules/#Modules","page":"Modules","title":"Modules","text":"","category":"section"},{"location":"man/modules/","page":"Modules","title":"Modules","text":"This project implements a number of ART-based models with options that modulate their behavior (see the options section of the Guide)","category":"page"},{"location":"man/modules/","page":"Modules","title":"Modules","text":"This page lists both the implemented models and some variants","category":"page"},{"location":"man/modules/#Implemented-Models","page":"Modules","title":"Implemented Models","text":"","category":"section"},{"location":"man/modules/","page":"Modules","title":"Modules","text":"This project has implementations of the following ART (unsupervised) and ARTMAP (supervised) modules:","category":"page"},{"location":"man/modules/","page":"Modules","title":"Modules","text":"ART\nFuzzyART: Fuzzy ART\nDVFA: Dual Vigilance Fuzzy ART\nDDVFA: Distributed Dual Vigilance Fuzzy ART\nARTMAP\nSFAM: Simplified Fuzzy ARTMAP\nFAM: Fuzzy ARTMAP\nDAM: Default ARTMAP","category":"page"},{"location":"man/modules/#Variants","page":"Modules","title":"Variants","text":"","category":"section"},{"location":"man/modules/","page":"Modules","title":"Modules","text":"Each module contains many options that modulate its behavior. Some of these options are used to modulate the internals of the module, such as switching the match and activation functions, to achieve different modules that are found in the literature.","category":"page"},{"location":"man/modules/","page":"Modules","title":"Modules","text":"These variants are:","category":"page"},{"location":"man/modules/","page":"Modules","title":"Modules","text":"Gamma-Normalized FuzzyART","category":"page"},{"location":"man/modules/#Gamma-Normalized-FuzzyART","page":"Modules","title":"Gamma-Normalized FuzzyART","text":"","category":"section"},{"location":"man/modules/","page":"Modules","title":"Modules","text":"A Gamma-Normalized FuzzyART is a FuzzyART module where the gamma normalization option is set on gamma_normalization=true and the kernel width parameter is set to gamma = 10 (gamma_ref is 1.0 by default):","category":"page"},{"location":"man/modules/","page":"Modules","title":"Modules","text":"my_gnfa = FuzzyART(gamma_normalization=true, gamma=5.0)","category":"page"},{"location":"man/modules/","page":"Modules","title":"Modules","text":"The gamma_normalization flag must be set high here because it also changes the thresholding value and match function of the module.","category":"page"},{"location":"getting-started/basic-example/#Basic-Example","page":"Basic Example","title":"Basic Example","text":"","category":"section"},{"location":"getting-started/basic-example/","page":"Basic Example","title":"Basic Example","text":"In the example below, we create a dataset generated from two multivariate Gaussian distributions in two dimensions, showing how an ART module can be used in unsupervised or simple supervised modes alongside an ARTMAP module that is explicitly supervised-only.","category":"page"},{"location":"getting-started/basic-example/","page":"Basic Example","title":"Basic Example","text":"# Copyright © 2021 Alexander L. Hayes\n# MIT License\n\nusing AdaptiveResonance\nusing Distributions, Random\nusing MLDataUtils\nusing Plots\n\n\"\"\"\nDemonstrates Unsupervised DDVFA, Supervised DDVFA, and (Supervised) SFAM on a toy problem\nwith two multivaria
[…]
