### 2021-10-31 AP6YC/AdaptiveResonance.jl#42 @markNZed
i wonder if setting the argument get_bmu to true would fix this?  https://github.com/AP6YC/AdaptiveResonance.jl/blob/f1227ea692643796ae3d38eb7267b0536f928688/src/ART/DDVFA.jl#L778
