### 2021-11-02 #3671 @adi3
@AP6YC please address the issues listed above that require your attention. Can you get these done within this month? I'd like to take this review forward.
