### 2021-11-02 #3671 @AP6YC
@adi3 Sure thing, end of the month is very feasible. I have been tackling the issues on the project's issues page, but I will be more diligent to give updates on them here too.
