### 2021-11-02 AP6YC/AdaptiveResonance.jl#41 @AP6YC
Good catch! I was hoping to fold in a lot of overlapping fixes with broader changes in a major release, so that is why things have stayed in `develop`.  I'll post an update here when that release is merged.
