### 2021-11-02 AP6YC/AdaptiveResonance.jl#42 @AP6YC
Good catch @hayesall, and @markNZed is correct. The actionable item is that it is something for me to document much better considering how important of a point it is.

During inference, ART algorithms are defined to return their category belief only if the match function vigilance test is passed. Because one may still want a "best-matching-unit" (i.e., highest activated category through the activation function) for situations where you want an answer regardless of mismatch, I added a `get_bmu` to the ART implementations.

What you see in the plot is the result of "default" ART behavior. I will certainly add a blurb in the usage section of the docs, but I think that the point might merit a full page with an example such as yours. I'll link the feature that implements it.
