### 2021-11-02 AP6YC/AdaptiveResonance.jl#25 @AP6YC
Good suggestions! I'll update the paper with a section reflecting exactly those bullet points.
