### 2021-11-03 1b4dcc4
Refactor SFAM, create common ARTMAP methods
