+++ src/ARTMAP/ARTMAP.jl
@@ -5,6 +5,7 @@ Description:
     Includes all of the ARTMAP (i.e., explicitly supervised) ART modules definitions.
 """
 
+include("common.jl")    # train!
 include("DAM.jl")       # Default ARTMAP
 include("FAM.jl")       # Fuzzy ARTMAP
 include("SFAM.jl")      # Simplified Fuzzy ARTMAP
