### 2021-11-03 d616094
Refactor DAM like SFAM
