### 2021-11-29 a44f624
Add target audience, theory, and dynamics to paper
