### 2021-11-29 AP6YC/AdaptiveResonance.jl#24 @AP6YC
PR #46: I updated the paper to include a target audience section that addresses your feedback. To tackle the types-of-problems idea, I wrote up a simplified summary of ART theory and algorithm basics to explain what ART is and what it is used for.
