### 2021-11-29 877f9d0
Add get_bmu sections to docs and README
