### 2021-11-29 AP6YC/AdaptiveResonance.jl#42 @AP6YC
PR #47 documents the mismatch vs. best-matching unit dynamics I mentioned above in the README and docs. My hope is that including it in the README's quickstart brings this usage to light quickly for the reader.
