### 2021-11-30 #3671 @AP6YC
You've got it; for starters, [PR 46](https://github.com/AP6YC/AdaptiveResonance.jl/pull/46) answers [Issue 24](https://github.com/AP6YC/AdaptiveResonance.jl/issues/24): the target audience within the statement of need.

These and other changes are in `develop` since switching to GitFlow, and I am hoping to make a major release once I have addressed all of the reviewers' concerns and recommendations.
