### 2021-11-30 AP6YC/AdaptiveResonance.jl#41 @AP6YC
PR #48 pulls all of the recent changes into the master, including a fix to this method definition issue. I have forgone from waiting for a major release to be able to address this issue and get other upcoming changes into master.
