### 2021-11-30 #3671 @AP6YC
[PR 48](https://github.com/AP6YC/AdaptiveResonance.jl/pull/48/) addresses [Issue 41](https://github.com/AP6YC/AdaptiveResonance.jl/issues/41): method redefinition (a bug, but listed here for completeness).
