### 2021-11-30 AP6YC/AdaptiveResonance.jl#26 @AP6YC
Thank you for mentioning those packages; I actually really like DemoCards, and it seems to use Literate under the hood to an extent. I think that that would cleanly organize the example page. I'll link the feature PR when I've got something to show for your approval.
