### 2021-11-30 AP6YC/AdaptiveResonance.jl#26 @AP6YC
PR #39 implemented the basic example page from a modification of your code above, but a separate upcoming PR will add the DemoCards examples.
