### 2021-11-30 #3671 @AP6YC
[PR 39](https://github.com/AP6YC/AdaptiveResonance.jl/pull/39) partially addresses [Issue 26](https://github.com/AP6YC/AdaptiveResonance.jl/issues/26): the "Hello World" example. An additional PR will come to clean up the `examples/` directory into the docs, which will also address [Issue 27](https://github.com/AP6YC/AdaptiveResonance.jl/issues/27): basic working examples.
