### 2021-11-30 eba9514
Add DemoCards dep to docs
