### 2021-11-30 AP6YC/AdaptiveResonance.jl#42 @AP6YC
I'll probably only close this once I get examples with DemoCards (Issues #26 and #27) that include showing the difference with and without `get_bmu`.
