+++ docs/demos/index.md
@@ -1,0 +1,19 @@
+# [Example Cards](@id examplecards)
+
+This section describes how you can set up your demos easily in _five lines of codes_.
+
+## Manage your demo files
+
+The rules of demo management are super simple:
+
+* you need one demo page (folder) to hold all the demo files
+* a demo page has several demo sections (subfolders)
+* a demo section either
+    * has other demo sections as nested subsections, or,
+    * has the demo files (`.md`, `.jl`)
+
+---
+
+{{{democards}}}
+
+---
