+++ docs/Project.toml
@@ -3,6 +3,7 @@ AdaptiveResonance = "3d72adc0-63d3-4141-bf9b-84450dd0395b"
 DemoCards = "311a05b2-6137-4a5a-b473-18580a3d38b5"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
 LiveServer = "16fef848-5104-11e9-1b77-fb7a48bbb589"
 MLDataUtils = "cc2ba9b6-d476-5e6d-8eaf-a92d5412d41d"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
