### 2021-12-01 5e034d7
Add basic ddvfa unsupervised example
