### 2021-12-01 8be54f6
Data util dev for examples
