### 2021-12-01 884b05a
Fix and refine ddvfa supervised example
