### 2021-12-01 42dd03b
Add sfam iris example
