+++ docs/examples/adaptiveresonance/incremental-batch.jl
@@ -1,0 +1,29 @@
+# ---
+# title: Incremental vs. Batch Example
+# id: incremental_batch
+# cover: ../assets/art.png
+# date: 2021-12-1
+# author: "[Sasha Petrenko](https://github.com/AP6YC)"
+# julia: 1.6
+# description: This demo illustrates how to use incremental training methods vs. batch training for all ART modules.
+# ---
+
+# All modules in `AdaptiveResonance.jl` are designed to handle incremental and batch training.
+# In fact, ART modules are generally incremental in their implementation, so their batch methods wrap the incremental ones and handle preprocessing, etc.
+
+# We begin with importing AdaptiveResonance for the ART modules and MLDatasets for some data utilities.
+using AdaptiveResonance # ART
+using MLDatasets        # Iris dataset
+using MLDataUtils       # Shuffling and splitting
+using Printf            # Formatted number printing
+
+# We will download the Iris dataset for its small size and benchmark use for clustering algorithms.
+Iris.download(i_accept_the_terms_of_use=true)
+features, labels = Iris.features(), Iris.labels()
+
+# Because the MLDatasets package gives us Iris labels as strings, we will use the `MLDataUtils.convertlabel` method with the `MLLabelUtils.LabelEnc.Indices` type to get a list of integers representing each class:
+labels = convertlabel(LabelEnc.Indices{Int}, labels)
+unique(labels)
+
+# Next, we will create a train/test split with the `MLDataUtils.stratifiedobs` utility:
+(X_train, y_train), (X_test, y_test) = stratifiedobs((features, labels))
