### 2021-12-01 5af8498
Remove boilerplate examples
