+++ versions.js
@@ -1,9 +1,10 @@
 var DOC_VERSIONS = [
   "stable",
+  "v0.4",
   "v0.3",
   "v0.2",
   "v0.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.3.7";
+var DOCUMENTER_NEWEST = "v0.4.0";
 var DOCUMENTER_STABLE = "stable";
