### 2021-12-02 6a6b131
Fix preprocessing in incremental vs batch, add tests
