### 2021-12-02 651e8a7
Add artmap asset to example card
