+++ docs/examples/adaptiveresonance/options.jl
@@ -1,0 +1,11 @@
+# ---
+# title: ART Options Example
+# id: options
+# cover: ../assets/art.png
+# date: 2021-12-2
+# author: "[Sasha Petrenko](https://github.com/AP6YC)"
+# julia: 1.6
+# description: This demo illustrates how to use options and modify the options for all ART and ARTMAP modules.
+# ---
+
+# TODO
