### 2021-12-03 b4539b2
Fix note, fix typo in demo description
