+++ docs/examples/adaptiveresonance/options.jl
@@ -8,4 +8,132 @@
 # description: This demo illustrates how to use options and modify the options for all ART and ARTMAP modules.
 # ---
 
-# TODO
+# The `AdaptiveResonance.jl` package has several ways of handling options for ART modules.
+# These methods are meant to give maximum flexibility to the user for sharing and interpreting options, which themselves vary between each module.
+
+# !!! note
+#     For more info on options in ART modules, see the guide in the docs on [ART options](@ref art_options).
+
+# First, we load our libraries
+using AdaptiveResonance
+
+# Every ART module has a default constructor, which can be instantiated in the usual way:
+
+my_fuzzyart = FuzzyART()
+typeof(my_fuzzyart)
+
+# Within every ART module is a [Parameters.jl](https://github.com/mauro3/Parameters.jl) struct named `opts` containing the options for the module
+
+## Check the FuzzyART options
+my_fuzzyart.opts
+
+# Note that the options here have the type `opts_FuzzyART`.
+# This nomenclature is used throughout the module to indicate an options type associated with an ART module.
+# For example, the options for a DDVFA module are `opts_DDVFA`:
+
+## Create a DDVFA module and check the type of the options
+my_ddvfa = DDVFA()
+typeof(my_ddvfa.opts)
+
+# In fact, we can create an instance of these options with a default constructor:
+
+## Create a separate options struct
+my_fuzzyart_opts = opts_FuzzyART()
+
+# In addition to the default constructor, we can construct ART modules by instantiating these options and passing them to the module during construction:
+
+## Instantiate an ART module by passing our options
+my_fuzzyart = FuzzyART(my_fuzzyart_opts)
+my_other_fuzzyart = FuzzyART(my_fuzzyart_opts)
+
+# Now to the good stuff: because of the behavior of the `Parameters.jl` type, each option has a default value that we can modify during instantiation with keyword arguments:
+
+## Change some of the default FuzzyART options
+my_fuzzyart_opts = opts_FuzzyART(
+    rho=0.6,
+    gamma_normalization=true
+)
+my_fuzzyart = FuzzyART(my_fuzzyart_opts)
+
+# As some syntactic sugar, we can pass these keyword arguments directly to the module during instantiation if we have no need to share option structs:
+
+## Pass these keyword arguments to the module directly
+my_fuzzyart = FuzzyART(
+    rho=0.6,
+    gamma_normalization=true
+)
+
+# !!! note
+#     All ART modules are designed to use this options struct internally when the parameters are needed.
+#     It is possible to change these parameters in the middle of training and evaluation, but some algorithmic instability may occur.
+
+# To see the effect that changing these parameters has on the modules, we can train and test them side-by-side.
+
+# We begin with importing AdaptiveResonance for the ART modules and MLDatasets for some data utilities.
+using MLDatasets        # Iris dataset
+using MLDataUtils       # Shuffling and splitting
+using Printf            # Formatted number printing
+using MultivariateStats # Principal component analysis (PCA)
+using Plots             # Plotting
+
+# We will download the Iris dataset for its small size and benchmark use for clustering algorithms.
+Iris.download(i_accept_the_terms_of_use=true)
+features, labels = Iris.features(), Iris.labels()
+
+# Because the MLDatasets package gives us Iris labels as strings, we will use the `MLDataUtils.convertlabel` method with the `MLLabelUtils.LabelEnc.Indices` type to get a list of integers representing each class:
[…]
