### 2021-12-03 7544895
Begin dataconfig example writing
