+++ docs/examples/adaptiveresonance/data_config.jl
@@ -1,0 +1,11 @@
+# ---
+# title: ART DataConfig Example
+# id: data_config
+# cover: ../assets/art.png
+# date: 2021-12-2
+# author: "[Sasha Petrenko](https://github.com/AP6YC)"
+# julia: 1.6
+# description: This demo illustrates how the data configuration object works for data preprocessing in ART modules that require it.
+# ---
+
+# ART
