### 2021-12-03 f9f92d2
Write data_config example
