+++ docs/examples/adaptiveresonance/data_config.jl
@@ -8,4 +8,62 @@
 # description: This demo illustrates how the data configuration object works for data preprocessing in ART modules that require it.
 # ---
 
-# ART
+# In their derivations, ART modules have some special requirements when it comes to their input features.
+# FuzzyART in particular, and subsequently its derivatives, has a requirement that the inputs be bounded and complement coded.
+# This is due to some consequences such as weight decay that occur when using real-valued patterns rather than binary ones (and hence operations like fuzzy membership).
+
+# Preprocessing of the features occurs as follows:
+# 1. The features are linearly normalized from 0 to 1 with respect to each feature with `linear_normalization`.
+# This is done according to some known bounds that each feature has.
+# 2. The features are then complement coded, meaning that the feature vector is appended to its 1-complement (i.e., x -> [x, 1-x]) with `complement_code`.
+
+# This preprocessing has the ultimate consequence that the input features must be bounded.
+# This many not be a problem in some offline applications with a fixed dataset, but in others where the bounds are not known, techniques such as sigmoidal limiting are often used to place an artificial limit.
+
+# Regardless, this process requires some *a-priori* knowledge about the minimums and maximums that each feature can have, which is stored as a preprocessing configuration.
+# This preprocessing configuration is saved in every ART module as a `DataConfig` object called `config`, which we can see is uninitialized at first:
+
+## Load the library
+using AdaptiveResonance
+
+## Create a new ART module and inspect its uninitialized data config `config`
+art = FuzzyART()
+art.config
+
+# In batch training mode, the minimums and maximums are detected automatically; the minimum and maximum values for every feature are saved and used for the preprocessing step at every subsequent iteration.
+
+## Load data
+using MLDatasets
+
+## We will download the Iris dataset for its small size and benchmark use for clustering algorithms.
+Iris.download(i_accept_the_terms_of_use=true)
+features, labels = Iris.features(), Iris.labels()
+
+## We will then train the FuzzyART module in unsupervised mode and see that the data config is now set
+y_hat_train = train!(art, features)
+art.config
+
+# !!! note
+#     This automatic detection of feature characteristics only occurs if the `config` is not already setup.
+#     If it is setup beforehand, then that config is used instead.
+
+# As mentioned before, we may not always have the luxury of having a representative dataset in advance.
+# Alternatively, we may know the bounds of the features but wish to run incrementally rather than in batch.
+# In these cases, we can setup the config the various `DataConfig` constructors.
+
+# For example, if the features are all bounded from -1 to 1, we have to also specify the original dimension of the data in `DataConfig(min, max, dim)`:
+
+## Reinitialize the FuzzyART module
+art = FuzzyART()
+## Tell the module that we have 20 features all ranging from -1 to 1
+art.config = DataConfig(-1, 1, 20)
+
+# If the features differ in their ranges, we can specify with `DataConfig(mins, maxs)`:
+
+## Assume some minimum and maximum values for each feature
+mins = [-1,-2,-1.5]
+maxs = [3, 2, 1]
+art.config = DataConfig(mins, max)
+
+# Here, we don't need to specify the feature dimensionality because it is inferred from the length of the range values.
+
