### 2021-12-03 1e91385
Add sections to examples, clean up doc errors
