### 2021-12-03 8b6cdd6
More examples docs cleanup
