### 2021-12-03 012b830
Retire old examples page and deprecated examples
