### 2021-12-03 AP6YC/AdaptiveResonance.jl#27 @AP6YC
PR #49 implements many examples now using the `DemoCards` package in the Examples section of the docs.
