### 2021-12-03 AP6YC/AdaptiveResonance.jl#42 @AP6YC
PR #49 implements those `DemoCards` examples in the Examples section of the docs.
