### 2021-12-03 AP6YC/AdaptiveResonance.jl#24 @AP6YC
I'll close this issue for now. If you guys find anything is unclear or lacking, I will gladly write up more 😊
