### 2021-12-03 #3671 @AP6YC
As promised, PR #49 now addresses Issues [26](https://github.com/AP6YC/AdaptiveResonance.jl/issues/26) [27](https://github.com/AP6YC/AdaptiveResonance.jl/issues/27) by implementing many working examples using the `DemoCards` package in the Examples section of the docs and removing the deprecated example page and folder.
