### 2021-12-03 f1f9268
Delete docs/examples/AdaptiveResonance directory
