+++ docs/examples/AdaptiveResonance/options.jl
@@ -1,158 +1,0 @@
-# ---
-# title: ART Options Example
-# id: options
-# cover: ../assets/art.png
-# date: 2021-12-2
-# author: "[Sasha Petrenko](https://github.com/AP6YC)"
-# julia: 1.6
-# description: This demo illustrates how to use options and modify the options for all ART and ARTMAP modules.
-# ---
-
-# ## Overview
-
-# The `AdaptiveResonance.jl` package has several ways of handling options for ART modules.
-# These methods are meant to give maximum flexibility to the user for sharing and interpreting options, which themselves vary between each module.
-
-# !!! note
-#     For more info on options in ART modules, see the guide in the docs on [ART options](@ref art_options).
-
-# ## ART Options
-
-# To get a feel for the ART options system, we will inspect different options and their instantiation methods.
-
-# ### Inspection
-
-# First, we load `AdaptiveResonance`:
-using AdaptiveResonance
-
-# Every ART module has a default constructor, which can be instantiated in the usual way:
-
-## Create a FuzzyART module with default options
-my_fuzzyart = FuzzyART()
-typeof(my_fuzzyart)
-
-# Within every ART module is a [Parameters.jl](https://github.com/mauro3/Parameters.jl) struct named `opts` containing the options for the module
-
-## Check the FuzzyART options
-my_fuzzyart.opts
-
-# Note that the options here have the type `opts_FuzzyART`.
-# This nomenclature is used throughout the module to indicate an options type associated with an ART module.
-# For example, the options for a DDVFA module are `opts_DDVFA`:
-
-## Create a DDVFA module and check the type of the options
-my_ddvfa = DDVFA()
-typeof(my_ddvfa.opts)
-
-# In fact, we can create an instance of these options with a default constructor:
-
-## Create a separate options struct
-my_fuzzyart_opts = opts_FuzzyART()
-
-# In addition to the default constructor, we can construct ART modules by instantiating these options and passing them to the module during construction:
-
-## Instantiate an ART module by passing our options
-my_fuzzyart = FuzzyART(my_fuzzyart_opts)
-my_other_fuzzyart = FuzzyART(my_fuzzyart_opts)
-
-# ### Specifying Options
-
-# Now to the good stuff: because of the behavior of the `Parameters.jl` type, each option has a default value that we can modify during instantiation with keyword arguments:
-
-## Change some of the default FuzzyART options
-my_fuzzyart_opts = opts_FuzzyART(
-    rho=0.6,
-    gamma_normalization=true
-)
-my_fuzzyart = FuzzyART(my_fuzzyart_opts)
-
-# As some syntactic sugar, we can pass these keyword arguments directly to the module during instantiation if we have no need to share option structs:
-
-## Pass these keyword arguments to the module directly
-my_fuzzyart = FuzzyART(
-    rho=0.6,
-    gamma_normalization=true
-)
-
-# Before training, we can also instantiate the model and alter the options afterward:
-my_fuzzyart = FuzzyART()
-my_fuzzyart.opts.rho=0.6
-
-# !!! note
-#     All ART modules are designed to use this options struct internally when the parameters are needed.
-#     It is possible to change these parameters in the middle of training and evaluation, but some algorithmic instability may occur.
-
-# ## Comparison
-
-# To see the effect that changing these parameters has on the modules, we can train and test them side-by-side.
-
-# We begin with importing AdaptiveResonance for the ART modules and MLDatasets for some data utilities.
-using MLDatasets        # Iris dataset
-using MLDataUtils       # Shuffling and splitting
[…]
