### 2021-12-03 92ad6d1
Add datadeps accept to docs build env
