### 2021-12-03 ea0a809
Up docs julia version, add JSON dep to make
