### 2021-12-03 9807c48
Test add scipy install
