+++ .github/workflows/Documentation.yml
@@ -24,6 +24,7 @@ jobs:
         with:
           version: '1.6'
       - name: Install dependencies
+        run: pip install scipy
         run: julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
       - name: Build and deploy
         env:
