+++ .github/workflows/Documentation.yml
@@ -24,8 +24,9 @@ jobs:
         with:
           version: '1.6'
       - name: Install dependencies
-        run: pip install scipy
-        run: julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
+        run: |
+          pip install scipy
+          julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
       - name: Build and deploy
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }} # For authentication with GitHub Actions token
