+++ docs/make.jl
@@ -1,7 +1,7 @@
 using Documenter
 using AdaptiveResonance
 using DemoCards
-using JSON
+# using JSON
 
 # Generate the demo files
 # this is the relative path to docs/
@@ -48,13 +48,13 @@ postprocess_cb()
 
 # a workdaround to github action that only push preview when PR has "push_preview" labels
 # issue: https://github.com/JuliaDocs/Documenter.jl/issues/1225
-function should_push_preview(event_path = get(ENV, "GITHUB_EVENT_PATH", nothing))
-    event_path === nothing && return false
-    event = JSON.parsefile(event_path)
-    haskey(event, "pull_request") || return false
-    labels = [x["name"] for x in event["pull_request"]["labels"]]
-    return "push_preview" in labels
- end
+# function should_push_preview(event_path = get(ENV, "GITHUB_EVENT_PATH", nothing))
+#     event_path === nothing && return false
+#     event = JSON.parsefile(event_path)
+#     haskey(event, "pull_request") || return false
+#     labels = [x["name"] for x in event["pull_request"]["labels"]]
+#     return "push_preview" in labels
+#  end
 
 deploydocs(
     repo="github.com/AP6YC/AdaptiveResonance.jl.git",
