+++ .github/workflows/Documentation.yml
@@ -22,7 +22,7 @@ jobs:
       - uses: actions/checkout@v2
       - uses: julia-actions/setup-julia@latest
         with:
-          version: '1.6'
+          version: '1.4'
       - name: Install dependencies
         run: |
           pip install scipy
