+++ dev/examples/art/ddvfa_supervised.jl
@@ -1,0 +1,39 @@
+using AdaptiveResonance # ART
+using MLDatasets        # Iris dataset
+using MLDataUtils       # Shuffling and splitting
+using Printf            # Formatted number printing
+
+Iris.download(i_accept_the_terms_of_use=true)
+features, labels = Iris.features(), Iris.labels()
+
+labels = convertlabel(LabelEnc.Indices{Int}, labels)
+unique(labels)
+
+(X_train, y_train), (X_test, y_test) = stratifiedobs((features, labels))
+
+art = DDVFA()
+
+# Train in simple supervised mode by passing the labels as a keyword argument.
+y_hat_train = train!(art, X_train, y=y_train)
+println("Training labels: ",  size(y_hat_train), " ", typeof(y_hat_train))
+
+# Classify both ways
+y_hat = AdaptiveResonance.classify(art, X_test)
+y_hat_bmu = AdaptiveResonance.classify(art, X_test, get_bmu=true)
+
+# Check the shape and type of the output labels
+println("Testing labels: ",  size(y_hat), " ", typeof(y_hat))
+println("Testing labels with bmu: ",  size(y_hat_bmu), " ", typeof(y_hat_bmu))
+
+# Calculate performance on training data, testing data, and with get_bmu
+perf_train = performance(y_hat_train, y_train)
+perf_test = performance(y_hat, y_test)
+perf_test_bmu = performance(y_hat_bmu, y_test)
+
+# Format each performance number for comparison
+@printf "Training performance: %.4f\n" perf_train
+@printf "Testing performance: %.4f\n" perf_test
+@printf "Best-matching unit testing performance: %.4f\n" perf_test_bmu
+
+# This file was generated using Literate.jl, https://github.com/fredrikekre/Literate.jl
+
