### 2021-12-04 AP6YC/AdaptiveResonance.jl#26 @AP6YC
PR #49 implements the promised DemoCards examples. Thanks again for the recommendations! There was a little bit of a learning curve and debugging on the CI end, but I've added proper DemoCards notebooks for each example using MLDatasets and MLDataUtils (inspired by your working example above).
