### 2021-12-06 #3671 @hayesall
Yes! I've checked off items as they've come in.

I'll review functionality again (there was some discussion on assumptions in https://github.com/AP6YC/AdaptiveResonance.jl/issues/42 that have now been clarified).
