### 2022-01-24 #3671 @aaronpeikert
@AP6YC  I am almost satisfied (as soon as you merge dev and formally release a new version). The only thing that prevents me from ticking a box is https://github.com/AP6YC/AdaptiveResonance.jl/issues/23

While I do not consider them blocking, https://github.com/AP6YC/AdaptiveResonance.jl/issues/28 & https://github.com/AP6YC/AdaptiveResonance.jl/issues/25 would really increase the value of the paper/software.
