### 2022-02-13 #3671 @adi3
@AP6YC can you address the final few issues brought up by @aaronpeikert ? I also see some of the issues opened by @hayesall as still open so please tend to them as well once you have a chance. We're on the home stretch here and need just a tiny more effort to get this across the finish line!
