### 2022-03-24 #3671 @aaronpeikert
@AP6YC Could you give us an update on this? Can we somehow support you in finishing the software paper?

As @adi3 says:

> We're on the home stretch here and need just a tiny more effort to get this across the finish line!
