### 2022-04-15 #3671 @AP6YC
All,

I sincerely apologize for my absence in this project and this thread, which despite life events I cannot excuse. I would like to thank @adi3 for reaching out to me via email to correct my neglect of this thread. To quickly conclude this review, and as a show of good faith to everyone's time, I have gone through this thread and listed out the remaining reviewers' concerns:

- [ ] https://github.com/AP6YC/AdaptiveResonance.jl/issues/23
- [ ] https://github.com/AP6YC/AdaptiveResonance.jl/issues/28
- [ ] https://github.com/AP6YC/AdaptiveResonance.jl/issues/25

In summary, I will make immediate changes to the paper itself to compare and contrast existing implementations of ART algorithms, and I will complete the work that I began before in formalizing the docstrings that document the various options for each ART module in the index.

My next message will be to show that these last concerns are completed for the reviewers.
