### 2022-04-22 #3671 @hayesall
Just about there! I've ticked off all my boxes.

A few things need fixed in the references, these may not be exhaustive:

---

![image](https://user-images.githubusercontent.com/11916674/164799656-5e030942-c010-4c96-be54-f081ac43a218.png)

- Adjust `author`

---

![image](https://user-images.githubusercontent.com/11916674/164799849-44716e3f-c2eb-4706-b25e-5a93cb0157c5.png)

- Issue number `(xxxx)` probably isn't correct.

---

![image](https://user-images.githubusercontent.com/11916674/164799552-0f8b878e-60ec-437a-abed-155f904deca0.png)

- Remove the space between `Code ?` in `title`

---

![image](https://user-images.githubusercontent.com/11916674/164798709-63b4e38d-73f4-4eb2-8099-4712aff1f953.png)

- Drop "Scholar's Mine" from `title`
- Add the DOI: https://doi.org/10.13140/RG.2.2.11823.25761

---

![image](https://user-images.githubusercontent.com/11916674/164799272-fd09ada9-14a1-4334-9acc-6c9d928f2c9d.png)

- `author` section needs adjusted

---

![image](https://user-images.githubusercontent.com/11916674/164799065-e6e3a7a0-a31e-42e9-a17a-57a2cea107c8.png)

- `(xxxx)` seems unlikely to be the correct issue number
