### 2022-04-22 #3671 @AP6YC
@hayesall All very good catches! Sorry for the inconsistencies in the bibliography. I addressed your concerns and more, such as software citations, with a recent hotfix into `master`.
