### 2022-04-29 #3671 @AP6YC
@adi3 I have archived the project at Zenodo. The release folding in all of the last changes has the following record, DOI, and version:
- Record: https://zenodo.org/record/6506781
- DOI: 10.5281/zenodo.6506781
- Version: v0.5.0

Does this work, or should I archive in another way?
