### 2022-05-04 #3671 @kyleniemeyer
Hi @AP6YC, I am the AEIC on duty this week doing some final checks before publishing.

I noticed that the author list of your Zenodo archive does not match the author list of the paper. Generally, these should match. Can you either clean up the author list in the Zenodo metadata or justify why these should be different?
