### 2022-05-04 #3671 @AP6YC
@kyleniemeyer Thank you for catching that! I suspect that the discrepancy exists because I folded in two PRs done by the two editors of the paper before the latest release, which pulled their names into the automatic Zenodo upload, and I overlooked adding another author to Zenodo in the first place.

I will clean up the upload metadata to correct the authors to reflect the paper. Will it be appropriate to move the two editors as contributors to the Zenodo upload, or should they be removed entirely?
