### 2022-05-04 #3671 @kyleniemeyer
@AP6YC if they did not contribute to the software itself, then I think it's most appropriate to just remove them entirely, but it's up to you
