### 2022-05-04 #3671 @AP6YC
@kyleniemeyer gotchya; I added them as "other" contributors since they did make software contributions. The author list for the upload is fixed now. Thanks again for pointing it out!
